Split morphic.js in multiples files.
