+++ package.json
@@ -1,0 +1,32 @@
+{
+  "name": "morphic.js",
+  "version": "1.0.0",
+  "description": "A lively Web-GUI inspired by Squeak",
+  "main": "morphic.js",
+  "dependencies": {},
+  "devDependencies": {
+    "gulp": "^3.8.10",
+    "gulp-concat": "^2.4.2",
+    "gulp-footer": "^1.0.5",
+    "gulp-header": "^1.2.2",
+    "gulp-rename": "^1.2.0"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jmoenig/morphic.js.git"
+  },
+  "keywords": [
+    "morphic",
+    "snap",
+    "scratch"
+  ],
+  "author": "Jens Mönig",
+  "license": "AGPL",
+  "bugs": {
+    "url": "https://github.com/jmoenig/morphic.js/issues"
+  },
+  "homepage": "https://github.com/jmoenig/morphic.js"
+}
