+++ src/rectangles.js
@@ -1,0 +1,311 @@
+// Rectangles //////////////////////////////////////////////////////////
+
+// Rectangle instance creation:
+
+function Rectangle(left, top, right, bottom) {
+    this.init(new Point((left || 0), (top || 0)),
+            new Point((right || 0), (bottom || 0)));
+}
+
+Rectangle.prototype.init = function (originPoint, cornerPoint) {
+    this.origin = originPoint;
+    this.corner = cornerPoint;
+};
+
+// Rectangle string representation: e.g. '[0@0 | 160@80]'
+
+Rectangle.prototype.toString = function () {
+    return '[' + this.origin.toString() + ' | ' +
+        this.extent().toString() + ']';
+};
+
+// Rectangle copying:
+
+Rectangle.prototype.copy = function () {
+    return new Rectangle(
+        this.left(),
+        this.top(),
+        this.right(),
+        this.bottom()
+    );
+};
+
+// creating Rectangle instances from Points:
+
+Point.prototype.corner = function (cornerPoint) {
+    // answer a new Rectangle
+    return new Rectangle(
+        this.x,
+        this.y,
+        cornerPoint.x,
+        cornerPoint.y
+    );
+};
+
+Point.prototype.rectangle = function (aPoint) {
+    // answer a new Rectangle
+    var org, crn;
+    org = this.min(aPoint);
+    crn = this.max(aPoint);
+    return new Rectangle(org.x, org.y, crn.x, crn.y);
+};
+
+Point.prototype.extent = function (aPoint) {
+    //answer a new Rectangle
+    var crn = this.add(aPoint);
+    return new Rectangle(this.x, this.y, crn.x, crn.y);
+};
+
+// Rectangle accessing - setting:
+
+Rectangle.prototype.setTo = function (left, top, right, bottom) {
+    // note: all inputs are optional and can be omitted
+
+    this.origin = new Point(
+        left || ((left === 0) ? 0 : this.left()),
+        top || ((top === 0) ? 0 : this.top())
+    );
+
+    this.corner = new Point(
+        right || ((right === 0) ? 0 : this.right()),
+        bottom || ((bottom === 0) ? 0 : this.bottom())
+    );
+};
+
+// Rectangle accessing - getting:
+
+Rectangle.prototype.area = function () {
+    //requires width() and height() to be defined
+    var w = this.width();
+    if (w < 0) {
+        return 0;
+    }
+    return Math.max(w * this.height(), 0);
+};
+
+Rectangle.prototype.bottom = function () {
+    return this.corner.y;
+};
+
+Rectangle.prototype.bottomCenter = function () {
+    return new Point(this.center().x, this.bottom());
+};
+
+Rectangle.prototype.bottomLeft = function () {
+    return new Point(this.origin.x, this.corner.y);
+};
+
+Rectangle.prototype.bottomRight = function () {
+    return this.corner.copy();
+};
+
+Rectangle.prototype.boundingBox = function () {
+    return this;
+};
+
+Rectangle.prototype.center = function () {
+    return this.origin.add(
+        this.corner.subtract(this.origin).floorDivideBy(2)
+    );
+};
+
+Rectangle.prototype.corners = function () {
+    return [this.origin,
+        this.bottomLeft(),
+        this.corner,
+        this.topRight()];
+};
+
+Rectangle.prototype.extent = function () {
+    return this.corner.subtract(this.origin);
+};
+
+Rectangle.prototype.height = function () {
+    return this.corner.y - this.origin.y;
+};
+
+Rectangle.prototype.left = function () {
+    return this.origin.x;
+};
+
+Rectangle.prototype.leftCenter = function () {
+    return new Point(this.left(), this.center().y);
+};
+
+Rectangle.prototype.right = function () {
+    return this.corner.x;
+};
+
+Rectangle.prototype.rightCenter = function () {
+    return new Point(this.right(), this.center().y);
+};
+
+Rectangle.prototype.top = function () {
+    return this.origin.y;
+};
+
+Rectangle.prototype.topCenter = function () {
+    return new Point(this.center().x, this.top());
+};
+
+Rectangle.prototype.topLeft = function () {
+    return this.origin;
+};
+
+Rectangle.prototype.topRight = function () {
+    return new Point(this.corner.x, this.origin.y);
+};
+
+Rectangle.prototype.width = function () {
+    return this.corner.x - this.origin.x;
+};
+
+Rectangle.prototype.position = function () {
+    return this.origin;
+};
+
+// Rectangle comparison:
+
+Rectangle.prototype.eq = function (aRect) {
+    return this.origin.eq(aRect.origin) &&
+        this.corner.eq(aRect.corner);
+};
+
+Rectangle.prototype.abs = function () {
+    var newOrigin, newCorner;
+
+    newOrigin = this.origin.abs();
+    newCorner = this.corner.max(newOrigin);
+    return newOrigin.corner(newCorner);
+};
+
+// Rectangle functions:
+
+Rectangle.prototype.insetBy = function (delta) {
+    // delta can be either a Point or a Number
+    var result = new Rectangle();
+    result.origin = this.origin.add(delta);
+    result.corner = this.corner.subtract(delta);
+    return result;
+};
+
+Rectangle.prototype.expandBy = function (delta) {
+    // delta can be either a Point or a Number
+    var result = new Rectangle();
+    result.origin = this.origin.subtract(delta);
+    result.corner = this.corner.add(delta);
+    return result;
+};
+
+Rectangle.prototype.growBy = function (delta) {
+    // delta can be either a Point or a Number
+    var result = new Rectangle();
+    result.origin = this.origin.copy();
+    result.corner = this.corner.add(delta);
+    return result;
+};
+
+Rectangle.prototype.intersect = function (aRect) {
+    var result = new Rectangle();
+    result.origin = this.origin.max(aRect.origin);
+    result.corner = this.corner.min(aRect.corner);
+    return result;
+};
+
+Rectangle.prototype.merge = function (aRect) {
+    var result = new Rectangle();
+    result.origin = this.origin.min(aRect.origin);
+    result.corner = this.corner.max(aRect.corner);
+    return result;
+};
+
+Rectangle.prototype.mergeWith = function (aRect) {
+    // mutates myself
+    this.origin = this.origin.min(aRect.origin);
+    this.corner = this.corner.max(aRect.corner);
+};
+
+Rectangle.prototype.round = function () {
+    return this.origin.round().corner(this.corner.round());
+};
+
+Rectangle.prototype.spread = function () {
+    // round me by applying floor() to my origin and ceil() to my corner
+    return this.origin.floor().corner(this.corner.ceil());
+};
+
+Rectangle.prototype.amountToTranslateWithin = function (aRect) {
+/*
+    Answer a Point, delta, such that self + delta is forced within
+    aRectangle. when all of me cannot be made to fit, prefer to keep
+    my topLeft inside. Taken from Squeak.
+*/
+    var dx = 0, dy = 0;
+
+    if (this.right() > aRect.right()) {
+        dx = aRect.right() - this.right();
+    }
+    if (this.bottom() > aRect.bottom()) {
+        dy = aRect.bottom() - this.bottom();
+    }
+    if ((this.left() + dx) < aRect.left()) {
+        dx = aRect.left() - this.left();
+    }
+    if ((this.top() + dy) < aRect.top()) {
+        dy = aRect.top() - this.top();
+    }
+    return new Point(dx, dy);
+};
+
+// Rectangle testing:
+
+Rectangle.prototype.containsPoint = function (aPoint) {
+    return this.origin.le(aPoint) && aPoint.lt(this.corner);
+};
+
+Rectangle.prototype.containsRectangle = function (aRect) {
+    return aRect.origin.gt(this.origin) &&
+        aRect.corner.lt(this.corner);
+};
+
+Rectangle.prototype.intersects = function (aRect) {
+    var ro = aRect.origin, rc = aRect.corner;
+    return (rc.x >= this.origin.x) &&
+        (rc.y >= this.origin.y) &&
+        (ro.x <= this.corner.x) &&
+        (ro.y <= this.corner.y);
+};
+
+Rectangle.prototype.isNearTo = function (aRect, threshold) {
+    var ro = aRect.origin, rc = aRect.corner, border = threshold || 0;
+    return (rc.x + border >= this.origin.x) &&
+        (rc.y  + border >= this.origin.y) &&
+        (ro.x - border <= this.corner.x) &&
+        (ro.y - border <= this.corner.y);
+};
+
+// Rectangle transforming:
+
+Rectangle.prototype.scaleBy = function (scale) {
+    // scale can be either a Point or a scalar
+    var o = this.origin.multiplyBy(scale),
+        c = this.corner.multiplyBy(scale);
+    return new Rectangle(o.x, o.y, c.x, c.y);
+};
+
+Rectangle.prototype.translateBy = function (factor) {
+    // factor can be either a Point or a scalar
+    var o = this.origin.add(factor),
+        c = this.corner.add(factor);
+    return new Rectangle(o.x, o.y, c.x, c.y);
+};
+
+// Rectangle converting:
+
+Rectangle.prototype.asArray = function () {
+    return [this.left(), this.top(), this.right(), this.bottom()];
+};
+
+Rectangle.prototype.asArray_xywh = function () {
+    return [this.left(), this.top(), this.width(), this.height()];
+};
