+++ src/todo.js
@@ -1,5489 +1,3 @@
-/*
-	morphic.js
-
-	a lively Web-GUI
-	inspired by Squeak
-
-	written by Jens Mönig
-	jens@moenig.org
-
-	Copyright (C) 2012 by Jens Mönig
-
-	this documentation last changed: April 07, 2013
-
-	This file is part of Snap!.
-
-	Snap! is free software: you can redistribute it and/or modify
-	it under the terms of the GNU Affero General Public License as
-	published by the Free Software Foundation, either version 3 of
-	the License, or (at your option) any later version.
-
-	This program is distributed in the hope that it will be useful,
-	but WITHOUT ANY WARRANTY; without even the implied warranty of
-	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	GNU Affero General Public License for more details.
-
-	You should have received a copy of the GNU Affero General Public License
-	along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-
-    documentation contents
-    ----------------------
-    I. inheritance hierarchy
-    II. object definition toc
-    III. yet to implement
-    IV. open issues
-    V. browser compatibility
-    VI. the big picture
-    VII. programming guide
-        (1) setting up a web page
-            (a) single world
-            (b) multiple worlds
-            (c) an application
-        (2) manipulating morphs
-        (3) events
-            (a) mouse events
-            (b) context menu
-            (c) dragging
-            (d) dropping
-            (e) keyboard events
-            (f) resize event
-            (g) combined mouse-keyboard events
-            (h) text editing events
-        (4) stepping
-        (5) creating new kinds of morphs
-        (6) development and user modes
-        (7) turtle graphics
-        (8) damage list housekeeping
-        (9) minifying morphic.js
-    VIII. acknowledgements
-    IX. contributors
-
-
-    I. hierarchy
-    -------------
-    the following tree lists all constructors hierarchically,
-    indentation indicating inheritance. Refer to this list to get a
-    contextual overview:
-
-    Color
-    Node
-        Morph
-            BlinkerMorph
-                CursorMorph
-            BouncerMorph*
-            BoxMorph
-                InspectorMorph
-                MenuMorph
-                MouseSensorMorph*
-                SpeechBubbleMorph
-            CircleBoxMorph
-                SliderButtonMorph
-                SliderMorph
-            ColorPaletteMorph
-                GrayPaletteMorph
-            ColorPickerMorph
-            FrameMorph
-                ScrollFrameMorph
-                    ListMorph
-                StringFieldMorph
-                WorldMorph
-            HandleMorph
-            HandMorph
-            PenMorph
-            ShadowMorph
-            StringMorph
-            TextMorph
-            TriggerMorph
-                MenuItemMorph
-    Point
-    Rectangle
-
-
-    II. toc
-    -------
-    the following list shows the order in which all constructors are
-    defined. Use this list to locate code in this document:
-
-
-    Global settings
-    Global functions
-
-    Color
-    Point
-    Rectangle
-    Node
-    Morph
-    ShadowMorph
-    HandleMorph
-    PenMorph
-    ColorPaletteMorph
-    GrayPaletteMorph
-    ColorPickerMorph
-    BlinkerMorph
-    CursorMorph
-    BoxMorph
-    SpeechBubbleMorph
-    CircleBoxMorph
-    SliderButtonMorph
-    SliderMorph
-    MouseSensorMorph*
-    InspectorMorph
-    MenuMorph
-    StringMorph
-    TextMorph
-    TriggerMorph
-    MenuItemMorph
-    FrameMorph
-    ScrollFrameMorph
-    ListMorph
-    StringFieldMorph
-    BouncerMorph*
-    HandMorph
-    WorldMorph
-
-    * included only for demo purposes
-
-
-    III. yet to implement
-    ---------------------
-    - keyboard support for scroll frames and lists
-    - virtual keyboard support for Android and IE
-
-
-    IV. open issues
-    ----------------
-    - blurry shadows don't work well in Chrome
-
-
-    V. browser compatibility
-    ------------------------
-    I have taken great care and considerable effort to make morphic.js
-    runnable and appearing exactly the same on all current browsers
-    available to me:
-
-    - Firefox for Windows
-    - Firefox for Mac
-    - Chrome for Windows (blurry shadows have some issues)
-    - Chrome for Mac
-    - Safari for Windows
-    - safari for Mac
-    - Safari for iOS (mobile)
-    - IE for Windows
-    - Opera for Windows
-    - Opera for Mac
-
-
-    VI. the big picture
-    -------------------
-    Morphic.js is completely based on Canvas and JavaScript, it is just
-    Morphic, nothing else. Morphic.js is very basic and covers only the
-    bare essentials:
-
-        * a stepping mechanism (a time-sharing multiplexer for lively
-          user interaction ontop of a single OS/browser thread)
-        * progressive display updates (only dirty rectangles are
-          redrawn in each display cycle)
-        * a tree structure
-        * a single World per Canvas element (although you can have
-          multiple worlds in multiple Canvas elements on the same web
-          page)
-        * a single Hand per World (but you can support multi-touch
-          events)
-        * a single text entry focus per World
-
-    In its current state morphic.js doesn't support Transforms (you
-    cannot rotate Morphs), but with PenMorph there already is a simple
-    LOGO-like turtle that you can use to draw onto any Morph it is
-    attached to. I'm planning to add special Morphs that support these
-    operations later on, but not for every Morph in the system.
-    Therefore these additions ("sprites" etc.) are likely to be part of
-    other libraries ("microworld.js") in separate files.
-
-    the purpose of morphic.js is to provide a malleable framework that
-    will let me experiment with lively GUIs for my hobby horse, which
-    is drag-and-drop, blocks based programming languages. Those things
-    (BYOB4 - http://byob.berkeley.edu) will be written using morphic.js
-    as a library.
-
-
-    VII. programming guide
-    ----------------------
-    Morphic.js provides a library for lively GUIs inside single HTML
-    Canvas elements. Each such canvas element functions as a "world" in
-    which other visible shapes ("morphs") can be positioned and
-    manipulated, often directly and interactively by the user. Morphs
-    are tree nodes and may contain any number of submorphs ("children").
-
-    All things visible in a morphic World are morphs themselves, i.e.
-    all text rendering, blinking cursors, entry fields, menus, buttons,
-    sliders, windows and dialog boxes etc. are created with morphic.js
-    rather than using HTML DOM elements, and as a consequence can be
-    changed and adjusted by the programmer regardless of proprietary
-    browser behavior.
-
-    Each World has an - invisible - "Hand" resembling the mouse cursor
-    (or the user's finger on touch screens) which handles mouse events,
-    and may also have a keyboardReceiver to handle key events.
-
-    The basic idea of Morphic is to continuously run display cycles and
-    to incrementally update the screen by only redrawing those  World
-    regions    which have been "dirtied" since the last redraw. Before
-    each shape is processed for redisplay it gets the chance to perform
-    a "step" procedure, thus allowing for an illusion of concurrency.
-
-
-    (1) setting up a web page
-    -------------------------
-    Setting up a web page for Morphic always involves three steps:
-    adding one or more Canvas elements, defining one or more worlds,
-    initializing and starting the main loop.
-
-
-    (a) single world
-    -----------------
-    Most commonly you will want your World to fill the browsers's whole
-    client area. This default situation is easiest and most straight
-    forward.
-
-    example html file:
-
-    <!DOCTYPE html>
-    <html>
-        <head>
-            <title>Morphic!</title>
-            <script type="text/javascript" src="morphic.js"></script>
-            <script type="text/javascript">
-                var world;
-
-                window.onload = function () {
-                    world = new WorldMorph(
-                        document.getElementById('world'));
-                    setInterval(loop, 50);
-                };
-
-                function loop() {
-                    world.doOneCycle();
-                }
-            </script>
-        </head>
-        <body>
-            <canvas id="world" tabindex="1" width="800" height="600">
-                <p>Your browser doesn't support canvas.</p>
-            </canvas>
-        </body>
-    </html>
-
-    if you use ScrollFrames or otherwise plan to support mouse wheel
-    scrolling events, you might also add the following inline-CSS
-    attribute to the Canvas element:
-
-        style="position: absolute;"
-
-    which will prevent the World to be scrolled around instead of the
-    elements inside of it in some browsers.
-
-
-    (b) multiple worlds
-    -------------------
-    If you wish to create a web page with more than one world, make
-    sure to prevent each world from auto-filling the whole page and
-    include    it in the main loop. It's also a good idea to give each
-    world its own tabindex:
-
-    example html file:
-
-    <!DOCTYPE html>
-    <html>
-        <head>
-            <title>Morphic!</title>
-            <script type="text/javascript" src="morphic.js"></script>
-            <script type="text/javascript">
-                var world1, world2;
-
-                window.onload = function () {
-                    world1 = new WorldMorph(
-                        document.getElementById('world1'), false);
-                    world2 = new WorldMorph(
-                        document.getElementById('world2'), false);
-                    setInterval(loop, 50);
-                };
-
-                function loop() {
-                    world1.doOneCycle();
-                    world2.doOneCycle();
-                }
-            </script>
-        </head>
-        <body>
-            <p>first world:</p>
-            <canvas id="world1" tabindex="1" width="600" height="400">
-                <p>Your browser doesn't support canvas.</p>
-            </canvas>
-            <p>second world:</p>
-            <canvas id="world2" tabindex="2" width="400" height="600">
-                <p>Your browser doesn't support canvas.</p>
-            </canvas>
-        </body>
-    </html>
-
-
-    (c) an application
-    -------------------
-    Of course, most of the time you don't want to just plain use the
-    standard Morhic World "as is" out of the box, but write your own
-    application (something like Scratch!) in it. For such an
-    application you'll create your own morph prototypes, perhaps
-    assemble your own "window frame" and bring it all to life in a
-    customized World state. the following example creates a simple
-    snake-like mouse drawing game.
-
-    example html file:
-
-    <!DOCTYPE html>
-    <html>
-        <head>
-            <title>touch me!</title>
-            <script type="text/javascript" src="morphic.js"></script>
-            <script type="text/javascript">
-                var worldCanvas, sensor;
-
-                window.onload = function () {
-                    var x, y, w, h;
-
-                    worldCanvas = document.getElementById('world');
-                    world = new WorldMorph(worldCanvas);
-                    world.isDevMode = false;
-                    world.color = new Color();
-
-                    w = 100;
-                    h = 100;
-
-                    x = 0;
-                    y = 0;
-
-                    while ((y * h) < world.height()) {
-                        while ((x * w) < world.width()) {
-                            sensor = new MouseSensorMorph();
-                            sensor.setPosition(new Point(x * w, y * h));
-                            sensor.alpha = 0;
-                            sensor.setExtent(new Point(w, h));
-                            world.add(sensor);
-                            x += 1;
-                        }
-                        x = 0;
-                        y += 1;
-                    }
-                    setInterval(loop, 50);
-                };
-
-                function loop() {
-                    world.doOneCycle();
-                }
-            </script>
-        </head>
-        <body bgcolor='black'>
-            <canvas id="world" width="800" height="600">
-                <p>Your browser doesn't support canvas.</p>
-            </canvas>
-        </body>
-    </html>
-
-    To get an idea how you can craft your own custom morph prototypes
-    I've included two examples which should give you an idea how to add
-    properties, override inherited methods and use the stepping
-    mechanism for "livelyness":
-
-        BouncerMorph
-        MouseSensorMorph
-
-    For the sake of sharing a single file I've included those examples
-    in morphic.js itself. Usually you'll define your additions in a
-    separate file and keep morphic.js untouched.
-
-
-    (2) manipulating morphs
-    -----------------------
-    There are many methods to programmatically manipulate morphs. Among
-    the most important and common ones among all morphs are the
-    following nine:
-
-    * hide()
-    * show()
-
-    * setPosition(aPoint)
-    * setExtent(aPoint)
-    * setColor(aColor)
-
-    * add(submorph)            - attaches submorph ontop
-    * addBack(submorph)        - attaches submorph underneath
-
-    * fullCopy()               - duplication
-    * destroy()                - deletion
-
-
-    (3) events
-    ----------
-    All user (and system) interaction is triggered by events, which are
-    passed on from the root element - the World - to its submorphs. The
-    World contains a list of system (browser) events it reacts to in its
-
-        initEventListeners()
-
-    method. Currently there are
-
-        - mouse
-        - drop
-        - keyboard
-        - (window) resize
-
-    events.
-
-    These system events are dispatched within the morphic World by the
-    World's Hand and its keyboardReceiver (usually the active text
-    cursor).
-
-
-    (a) mouse events:
-    -----------------
-    The Hand dispatches the following mouse events to relevant morphs:
-
-        mouseDownLeft
-        mouseDownRight
-        mouseClickLeft
-        mouseClickRight
-        mouseDoubleClick
-        mouseEnter
-        mouseLeave
-        mouseEnterDragging
-        mouseLeaveDragging
-        mouseMove
-        mouseScroll
-
-    If you wish your morph to react to any such event, simply add a
-    method of the same name as the event, e.g:
-
-        MyMorph.prototype.mouseMove = function(pos) {};
-
-    The only optional parameter of such a method is a Point object
-    indicating the current position of the Hand inside the World's
-    coordinate system.
-
-    Events may be "bubbled" up a morph's owner chain by calling
-
-        this.escalateEvent(functionName, arg)
-
-    in the event handler method's code.
-
-    Likewise, removing the event handler method will render your morph
-    passive to the event in question.
-
-
-    (b) context menu:
-    -----------------
-    By default right-clicking (or single-finger tap-and-hold) on a morph
-    also invokes its context menu (in addition to firing the
-    mouseClickRight event). A morph's context menu can be customized by
-    assigning a Menu instance to its
-
-        customContextMenu
-
-    property, or altogether suppressed by overriding its inherited
-
-        contextMenu()
-
-    method.
-
-
-    (c) dragging:
-    -------------
-    Dragging a morph is initiated when the left mouse button is pressed,
-    held and the mouse is moved.
-
-    You can control whether a morph is draggable by setting its
-
-        isDraggable
-
-    property either to false or true. If a morph isn't draggable itself
-    it will pass the pick-up request up its owner chain. This lets you
-    create draggable composite morphs like Windows, DialogBoxes,
-    Sliders etc.
-
-    Sometimes it is desireable to make "template" shapes which cannot be
-    moved themselves, but from which instead duplicates can be peeled
-    off. This is especially useful for building blocks in construction
-    kits, e.g. the MIT-Scratch palette. Morphic.js lets you control this
-    functionality by setting the
-
-        isTemplate
-
-    property flag to true for any morph whose "isDraggable" property is
-    turned off. When dragging such a Morph the hand will instead grab
-    a duplicate of the template whose "isDraggable" flag is true and
-    whose "isTemplate" flag is false, in other words: a non-template.
-
-    Dragging is indicated by adding a drop shadow to the morph in hand.
-    If a morph follows the hand without displaying a drop shadow it is
-    merely being moved about without changing its parent (owner morph),
-    e.g. when "dragging" a morph handle to resize its owner, or when
-    "dragging" a slider button.
-
-    Right before a morph is picked up its
-
-        prepareToBeGrabbed(handMorph)
-
-    method is invoked, if it is present. Immediately after the pick-up
-    the former parent's
-
-        reactToGrabOf(grabbedMorph)
-
-    method is called, again only if it exists.
-
-    Similar to events, these  methods are optional and don't exist by
-    default. For a simple example of how they can be used to adjust
-    scroll bars in a scroll frame please have a look at their
-    implementation in FrameMorph.
-
-
-    (d) dropping:
-    -------------
-    Dropping is triggered when the left mouse button is either pressed
-    or released while the Hand is dragging a morph.
-
-    Dropping a morph causes it to become embedded in a new owner morph.
-    You can control this embedding behavior by setting the prospective
-    drop target's
-
-        acceptsDrops
-
-    property to either true or false, or by overriding its inherited
-
-        wantsDropOf(aMorph)
-
-    method.
-
-    Right after a morph has been dropped its
-
-        justDropped(handMorph)
-
-    method is called, and its new parent's
-
-        reactToDropOf(droppedMorph, handMorph)
-
-    method is invoked, again only if each method exists.
-
-    Similar to events, these  methods are optional and by default are
-    not present in morphs by default (watch out for inheritance,
-    though!). For a simple example of how they can be used to adjust
-    scroll bars in a scroll frame please have a look at their
-    implementation in FrameMorph.
-
-    Drops of image elements from outside the world canvas are dispatched as
-
-        droppedImage(aCanvas, name)
-        droppedSVG(anImage, name)
-
-    events to interested Morphs at the mouse pointer. If you want you Morph
-    to e.g. import outside images you can add the droppedImage() and / or the
-    droppedSVG() methods to it. The parameter passed to the event handles is
-    a new offscreen canvas element representing a copy of the original image
-    element which can be directly used, e.g. by assigning it to another
-    Morph's image property. In the case of a dropped SVG it is an image
-    element (not a canvas), which has to be rasterized onto a canvas before
-    it can be used. The benefit of handling SVGs as image elements is that
-    rasterization can be deferred until the destination scale is known, taking
-    advantage of SVG's ability for smooth scaling. If instead SVGs are to be
-    rasterized right away, you can set the
-
-        MorphicPreferences.rasterizeSVGs
-
-    preference to <true>. In this case dropped SVGs also trigger the
-    droppedImage() event with a canvas containing a rasterized version of the
-    SVG.
-
-    The same applies to drops of audio or text files from outside the world
-    canvas.
-
-    Those are dispatched as
-
-        droppedAudio(anAudio, name)
-        droppedText(aString, name)
-
-    events to interested Morphs at the mouse pointer.
-
-    if none of the above content types can be determined, the file contents
-    is dispatched as an ArrayBuffer to interested Morphs:
-
-        droppedBinary(anArrayBuffer, name)
-
-
-    (e) keyboard events
-    -------------------
-    The World dispatches the following key events to its active
-    keyboardReceiver:
-
-        keypress
-        keydown
-        keyup
-
-    Currently the only morph which acts as keyboard receiver is
-    CursorMorph, the basic text editing widget. If you wish to add
-    keyboard support to your morph you need to add event handling
-    methods for
-
-        processKeyPress(event)
-        processKeyDown(event)
-        processKeyUp(event)
-
-    and activate them by assigning your morph to the World's
-
-        keyboardReceiver
-
-    property.
-
-    Note that processKeyUp() is optional and doesn't have to be present
-    if your morph doesn't require it.
-
-
-    (f) resize event
-    ----------------
-    The Window resize event is handled by the World and allows the
-    World's extent to be adjusted so that it always completely fills
-    the browser's visible page. You can turn off this default behavior
-    by setting the World's
-
-        useFillPage
-
-    property to false.
-
-    Alternatively you can also initialize the World with the
-    useFillPage switch turned off from the beginning by passing the
-    false value as second parameter to the World's constructor:
-
-        world = new World(aCanvas, false);
-
-    Use this when creating a web page with multiple Worlds.
-
-    if "useFillPage" is turned on the World dispatches an
-
-        reactToWorldResize(newBounds)
-
-    events to all of its children (toplevel only), allowing each to
-    adjust to the new World bounds by implementing a corresponding
-    method, the passed argument being the World's new dimensions after
-    completing the resize. By default, the "reactToWorldResize" Method
-    does not exist.
-
-    Example:
-
-    Add the following method to your Morph to let it automatically
-    fill the whole World, but leave a 10 pixel border uncovered:
-
-        MyMorph.prototype.reactToWorldResize = function (rect) {
-            this.changed();
-            this.bounds = rect.insetBy(10);
-            this.drawNew();
-            this.changed();
-        };
-
-
-    (g) combined mouse-keyboard events
-    ----------------------------------
-    Occasionally you'll want an object to react differently to a mouse
-    click or to some other mouse event while the user holds down a key
-    on the keyboard. Such "shift-click", "ctl-click", or "alt-click"
-    events can be implemented by querying the World's
-
-        currentKey
-
-    property inside the function that reacts to the mouse event. This
-    property stores the keyCode of the key that's currently pressed.
-    Once the key is released by the user it reverts to null.
-
-
-    (h) text editing events
-    -----------------------
-    Much of Morphic's "liveliness" comes out of allowing text elements
-    (instances of either single-lined StringMorph or multi-lined TextMorph)
-    to be directly manipulated and edited by users. This requires other
-    objects which may have an interest in the text element's state to react
-    appropriately. Therefore text elements and their manipulators emit
-    a stream of events, mostly by "bubbling" them up the text element's
-    owner chain. Text elements' parents are notified about the following
-    events:
-
-    Whenever the user presses a key on the keyboard while a text element
-    is being edited, a
-
-        reactToKeystroke(event)
-
-    is escalated up its parent chain, the "event" parameter being the
-    original one received by the World.
-
-    Once the user has completed the edit, the following events are
-    dispatched:
-
-        accept() - <enter> was pressed on a single line of text
-        cancel() - <esc> was pressed on any text element
-
-    Note that "accept" only gets triggered by single-line texte elements,
-    as the <enter> key is used to insert line breaks in multi-line
-    elements. Therefore, whenever a text edit is terminated by the user
-    (accepted, cancelled or otherwise),
-
-        reactToEdit(StringOrTextMorph)
-
-    is triggered.
-
-    If the MorphicPreference's
-
-        useSliderForInput
-
-    setting is turned on, a slider is popped up underneath the currently
-    edited text element letting the user insert numbers out of the given
-    slider range. Whenever this happens, i.e. whenever the slider is moved
-    or while the slider button is pressed, a stream of
-
-        reactToSliderEdit(StringOrTextMorph)
-
-    events is dispatched, allowing for "Bret-Victor" style "live coding"
-    applications.
-
-    In addition to user-initiated events text elements also emit
-    change notifications to their direct parents whenever their drawNew()
-    method is invoked. That way complex Morphs containing text elements
-    get a chance to react if something about the embedded text has been
-    modified programmatically. These events are:
-
-        layoutChanged() - sent from instances of TextMorph
-        fixLayout() - sent from instances of StringMorph
-
-    they are different so that Morphs which contain both multi-line and
-    single-line text elements can hold them apart.
-
-
-    (4) stepping
-    ------------
-    Stepping is what makes Morphic "magical". Two properties control
-    a morph's stepping behavior: the fps attribute and the step()
-    method.
-
-    By default the
-
-        step()
-
-    method does nothing. As you can see in the examples of BouncerMorph
-    and MouseSensorMorph you can easily override this inherited method
-    to suit your needs.
-
-    By default the step() method is called once per display cycle.
-    Depending on the number of actively stepping morphs and the
-    complexity of your step() methods this can cause quite a strain on
-    your CPU, and also result in your application behaving differently
-    on slower computers than on fast ones.
-
-    setting
-
-        myMorph.fps
-
-    to a number lower than the interval for the main loop lets you free
-    system resources (albeit at the cost of a less responsive or slower
-    behavior for this particular morph).
-
-
-    (5) creating new kinds of morphs
-    --------------------------------
-    The real fun begins when you start to create new kinds of morphs
-    with customized shapes. Imagine, e.g. jigsaw puzzle pieces or
-    musical notes. For this you have to override the default
-
-        drawNew()
-
-    method.
-
-    This method creates a new offscreen Canvas and stores it in
-    the morph's
-
-        image
-
-    property.
-
-    Use the following template for a start:
-
-        MyMorph.prototype.drawNew = function() {
-            var context;
-            this.image = newCanvas(this.extent());
-            context = this.image.getContext('2d');
-            // use context to paint stuff here
-        };
-
-    If your new morph stores or references other morphs outside of the
-    submorph tree in other properties, be sure to also override the
-    default
-
-        copyRecordingReferences()
-
-    method accordingly if you want it to support duplication.
-
-
-    (6) development and user modes
-    ------------------------------
-    When working with Squeak on Scratch or BYOB among the features I
-    like the best and use the most is inspecting what's going on in
-    the World while it is up and running. That's what development mode
-    is for (you could also call it debug mode). In essence development
-    mode controls which context menu shows up. In user mode right
-    clicking (or double finger tapping) a morph invokes its
-
-        customContextMenu
-
-    property, whereas in development mode only the general
-
-        developersMenu()
-
-    method is called and the resulting menu invoked. The developers'
-    menu features Gui-Builder-wise functionality to directly inspect,
-    take apart, reassamble and otherwise manipulate morphs and their
-    contents.
-
-    Instead of using the "customContextMenu" property you can also
-    assign a more dynamic contextMenu by overriding the general
-
-        userMenu()
-
-    method with a customized menu constructor. The difference between
-    the customContextMenu property and the userMenu() method is that
-    the former is also present in development mode and overrides the
-    developersMenu() result. For an example of how to use the
-    customContextMenu property have a look at TextMorph's evaluation
-    menu, which is used for the Inspector's evaluation pane.
-
-    When in development mode you can inspect every Morph's properties
-    with the inspector, including all of its methods. The inspector
-    also lets you add, remove and rename properties, and even edit
-    their values at runtime. Like in a Smalltalk environment the inspect
-    features an evaluation pane into which you can type in arbitrary
-    JavaScript code and evaluate it in the context of the inspectee.
-
-    Use switching between user and development modes while you are
-    developing an application and disable switching to development once
-    you're done and deploying, because generally you don't want to
-    confuse end-users with inspectors and meta-level stuff.
-
-
-    (7) turtle graphics
-    -------------------
-
-    The basic Morphic kernel features a simple LOGO turtle constructor
-    called
-
-        PenMorph
-
-    which you can use to draw onto its parent Morph. By default every
-    Morph in the system (including the World) is able to act as turtle
-    canvas and can display pen trails. Pen trails will be lost whenever
-    the trails morph (the pen's parent) performs a "drawNew()"
-    operation. If you want to create your own pen trails canvas, you
-    may wish to modify its
-
-        penTrails()
-
-    property, so that it keeps a separate offscreen canvas for pen
-    trails (and doesn't loose these on redraw).
-
-    the following properties of PenMorph are relevant for turtle
-    graphics:
-
-        color        - a Color
-        size        - line width of pen trails
-        heading        - degrees
-        isDown        - drawing state
-
-    the following commands can be used to actually draw something:
-
-        up()        - lift the pen up, further movements leave no trails
-        down()        - set down, further movements leave trails
-        clear()        - remove all trails from the current parent
-        forward(n)    - move n steps in the current direction (heading)
-        turn(n)        - turn right n degrees
-
-    Turtle graphics can best be explored interactively by creating a
-    new PenMorph object and by manipulating it with the inspector
-    widget.
-
-    NOTE: PenMorph has a special optimization for recursive operations
-    called
-
-        warp(function)
-
-    You can significantly speed up recursive ops and increase the depth
-    of recursion that's displayable by wrapping WARP around your
-    recursive function call:
-
-    example:
-
-        myPen.warp(function () {
-            myPen.tree(12, 120, 20);
-        })
-
-    will be much faster than just invoking the tree function, because it
-    prevents the parent's parent from keeping track of every single line
-    segment and instead redraws the outcome in a single pass.
-
-
-    (8) damage list housekeeping
-    ----------------------------
-    Morphic's progressive display update comes at the cost of having to
-    cycle through a list of "broken rectangles" every display cycle. If
-    this list gets very long working this damage list can lead to a
-    seemingly dramatic slow-down of the Morphic system. Typically this
-    occurs when updating the layout of complex Morphs with very many
-    submorphs, e.g. when resizing an inspector window.
-
-    An effective strategy to cope with this is to use the inherited
-
-        trackChanges
-
-    property of the Morph prototype for damage list housekeeping.
-
-    The trackChanges property of the Morph prototype is a Boolean switch
-    that determines whether the World's damage list ('broken' rectangles)
-    tracks changes. By default the switch is always on. If set to false
-    changes are not stored. This can be very useful for housekeeping of
-    the damage list in situations where a large number of (sub-) morphs
-    are changed more or less at once. Instead of keeping track of every
-    single submorph's changes tremendous performance improvements can be
-    achieved by setting the trackChanges flag to false before propagating
-    the layout changes, setting it to true again and then storing the full
-    bounds of the surrounding morph. An an example refer to the
-
-        moveBy()
-
-    method of HandMorph, and to the
-
-        fixLayout()
-
-    method of InspectorMorph, or the
-
-        startLayout()
-        endLayout()
-
-    methods of SyntaxElementMorph in the Snap application.
-
-
-    (9) minifying morphic.js
-    ------------------------
-    Coming from Smalltalk and being a Squeaker at heart I am a huge fan
-    of browsing the code itself to make sense of it. Therefore I have
-    included this documentation and (too little) inline comments so all
-    you need to get going is this very file.
-
-    Nowadays with live streaming HD video even on mobile phones 250 KB
-    shouldn't be a big strain on bandwith, still minifying and even
-    compressing morphic.js down do about 100 KB may sometimes improve
-    performance in production use.
-
-    Being an attorney-at-law myself you programmer folk keep harassing
-    me with rabulistic nitpickings about free software licenses. I'm
-    releasing morphic.js under an AGPL license. Therefore please make
-    sure to adhere to that license in any minified or compressed version.
-
-
-    VIII. acknowledgements
-    ----------------------
-    The original Morphic was designed and written by Randy Smith and
-    John Maloney for the SELF programming language, and later ported to
-    Squeak (Smalltalk) by John Maloney and Dan Ingalls, who has also
-    ported it to JavaScript (the Lively Kernel), once again setting
-    a "Gold Standard" for self sustaining systems which morphic.js
-    cannot and does not aspire to meet.
-
-    This Morphic implementation for JavaScript is not a direct port of
-    Squeak's Morphic, but still many individual functions have been
-    ported almost literally from Squeak, sometimes even including their
-    comments, e.g. the morph duplication mechanism fullCopy(). Squeak
-    has been a treasure trove, and if morphic.js looks, feels and
-    smells a lot like Squeak, I'll take it as a compliment.
-
-    Evelyn Eastmond has inspired and encouraged me with her wonderful
-    implementation of DesignBlocksJS. Thanks for sharing code, ideas
-    and enthusiasm for programming.
-
-    John Maloney has been my mentor and my source of inspiration for
-    these Morphic experiments. Thanks for the critique, the suggestions
-    and explanations for all things Morphic and for being my all time
-    programming hero.
-
-    I have originally written morphic.js in Florian Balmer's Notepad2
-    editor for Windows and later switched to Apple's Dashcode. I've also
-    come to depend on both Douglas Crockford's JSLint, Mozilla's Firebug
-    and Google's Chrome to get it right.
-
-
-    IX. contributors
-    ----------------------
-    Joe Otto found and fixed many early bugs and taught me some tricks.
-    Nathan Dinsmore contributed mouse wheel scrolling, cached
-    background texture handling and countless bug fixes.
-    Ian Reynolds contributed backspace key handling for Chrome.
-    Davide Della Casa contributed performance optimizations for Firefox.
-
-    - Jens Mönig
-*/
-// Global settings /////////////////////////////////////////////////////
-
-/*global window, HTMLCanvasElement, getMinimumFontHeight, FileReader, Audio,
-FileList, getBlurredShadowSupport*/
-
-var morphicVersion = '2014-November-20';
-var modules = {}; // keep track of additional loaded modules
-var useBlurredShadows = getBlurredShadowSupport(); // check for Chrome-bug
-
-var standardSettings = {
-    minimumFontHeight: getMinimumFontHeight(), // browser settings
-    globalFontFamily: '',
-    menuFontName: 'sans-serif',
-    menuFontSize: 12,
-    bubbleHelpFontSize: 10,
-    prompterFontName: 'sans-serif',
-    prompterFontSize: 12,
-    prompterSliderSize: 10,
-    handleSize: 15,
-    scrollBarSize: 12,
-    mouseScrollAmount: 40,
-    useSliderForInput: false,
-    useVirtualKeyboard: true,
-    isTouchDevice: false, // turned on by touch events, don't set
-    rasterizeSVGs: false,
-    isFlat: false
-};
-
-var touchScreenSettings = {
-    minimumFontHeight: standardSettings.minimumFontHeight,
-    globalFontFamily: '',
-    menuFontName: 'sans-serif',
-    menuFontSize: 24,
-    bubbleHelpFontSize: 18,
-    prompterFontName: 'sans-serif',
-    prompterFontSize: 24,
-    prompterSliderSize: 20,
-    handleSize: 26,
-    scrollBarSize: 24,
-    mouseScrollAmount: 40,
-    useSliderForInput: true,
-    useVirtualKeyboard: true,
-    isTouchDevice: false,
-    rasterizeSVGs: false,
-    isFlat: false
-};
-
-var MorphicPreferences = standardSettings;
-
-// Global Functions ////////////////////////////////////////////////////
-
-function nop() {
-    // do explicitly nothing
-    return null;
-}
-
-function localize(string) {
-    // override this function with custom localizations
-    return string;
-}
-
-function isNil(thing) {
-    return thing === undefined || thing === null;
-}
-
-function contains(list, element) {
-    // answer true if element is a member of list
-    return list.some(function (any) {
-        return any === element;
-    });
-}
-
-function detect(list, predicate) {
-    // answer the first element of list for which predicate evaluates
-    // true, otherwise answer null
-    var i, size = list.length;
-    for (i = 0; i < size; i += 1) {
-        if (predicate.call(null, list[i])) {
-            return list[i];
-        }
-    }
-    return null;
-}
-
-function sizeOf(object) {
-    // answer the number of own properties
-    var size = 0, key;
-    for (key in object) {
-        if (Object.prototype.hasOwnProperty.call(object, key)) {
-            size += 1;
-        }
-    }
-    return size;
-}
-
-function isString(target) {
-    return typeof target === 'string' || target instanceof String;
-}
-
-function isObject(target) {
-    return target !== null &&
-        (typeof target === 'object' || target instanceof Object);
-}
-
-function radians(degrees) {
-    return degrees * Math.PI / 180;
-}
-
-function degrees(radians) {
-    return radians * 180 / Math.PI;
-}
-
-function fontHeight(height) {
-    var minHeight = Math.max(height, MorphicPreferences.minimumFontHeight);
-    return minHeight * 1.2; // assuming 1/5 font size for ascenders
-}
-
-function newCanvas(extentPoint) {
-    // answer a new empty instance of Canvas, don't display anywhere
-    var canvas, ext;
-    ext = extentPoint || {x: 0, y: 0};
-    canvas = document.createElement('canvas');
-    canvas.width = ext.x;
-    canvas.height = ext.y;
-    return canvas;
-}
-
-function getMinimumFontHeight() {
-    // answer the height of the smallest font renderable in pixels
-    var str = 'I',
-        size = 50,
-        canvas = document.createElement('canvas'),
-        ctx,
-        maxX,
-        data,
-        x,
-        y;
-    canvas.width = size;
-    canvas.height = size;
-    ctx = canvas.getContext('2d');
-    ctx.font = '1px serif';
-    maxX = ctx.measureText(str).width;
-    ctx.fillStyle = 'black';
-    ctx.textBaseline = 'bottom';
-    ctx.fillText(str, 0, size);
-    for (y = 0; y < size; y += 1) {
-        for (x = 0; x < maxX; x += 1) {
-            data = ctx.getImageData(x, y, 1, 1);
-            if (data.data[3] !== 0) {
-                return size - y + 1;
-            }
-        }
-    }
-    return 0;
-}
-
-function getBlurredShadowSupport() {
-    // check for Chrome issue 90001
-    // http://code.google.com/p/chromium/issues/detail?id=90001
-    var source, target, ctx;
-    source = document.createElement('canvas');
-    source.width = 10;
-    source.height = 10;
-    ctx = source.getContext('2d');
-    ctx.fillStyle = 'rgb(255, 0, 0)';
-    ctx.beginPath();
-    ctx.arc(5, 5, 5, 0, Math.PI * 2, true);
-    ctx.closePath();
-    ctx.fill();
-    target = document.createElement('canvas');
-    target.width = 10;
-    target.height = 10;
-    ctx = target.getContext('2d');
-    ctx.shadowBlur = 10;
-    ctx.shadowColor = 'rgba(0, 0, 255, 1)';
-    ctx.drawImage(source, 0, 0);
-    return ctx.getImageData(0, 0, 1, 1).data[3] ? true : false;
-}
-
-function getDocumentPositionOf(aDOMelement) {
-    // answer the absolute coordinates of a DOM element in the document
-    var pos, offsetParent;
-    if (aDOMelement === null) {
-        return {x: 0, y: 0};
-    }
-    pos = {x: aDOMelement.offsetLeft, y: aDOMelement.offsetTop};
-    offsetParent = aDOMelement.offsetParent;
-    while (offsetParent !== null) {
-        pos.x += offsetParent.offsetLeft;
-        pos.y += offsetParent.offsetTop;
-        if (offsetParent !== document.body &&
-                offsetParent !== document.documentElement) {
-            pos.x -= offsetParent.scrollLeft;
-            pos.y -= offsetParent.scrollTop;
-        }
-        offsetParent = offsetParent.offsetParent;
-    }
-    return pos;
-}
-
-function clone(target) {
-    // answer a new instance of target's type
-    if (typeof target === 'object') {
-        var Clone = function () {nop(); };
-        Clone.prototype = target;
-        return new Clone();
-    }
-    return target;
-}
-
-function copy(target) {
-    // answer a shallow copy of target
-    var value, c, property;
-
-    if (typeof target !== 'object') {
-        return target;
-    }
-    value = target.valueOf();
-    if (target !== value) {
-        return new target.constructor(value);
-    }
-    if (target instanceof target.constructor &&
-            target.constructor !== Object) {
-        c = clone(target.constructor.prototype);
-        for (property in target) {
-            if (Object.prototype.hasOwnProperty.call(target, property)) {
-                c[property] = target[property];
-            }
-        }
-    } else {
-        c = {};
-        for (property in target) {
-            if (!c[property]) {
-                c[property] = target[property];
-            }
-        }
-    }
-    return c;
-}
-
-// Colors //////////////////////////////////////////////////////////////
-
-// Color instance creation:
-
-function Color(r, g, b, a) {
-    // all values are optional, just (r, g, b) is fine
-    this.r = r || 0;
-    this.g = g || 0;
-    this.b = b || 0;
-    this.a = a || ((a === 0) ? 0 : 1);
-}
-
-// Color string representation: e.g. 'rgba(255,165,0,1)'
-
-Color.prototype.toString = function () {
-    return 'rgba(' +
-        Math.round(this.r) + ',' +
-        Math.round(this.g) + ',' +
-        Math.round(this.b) + ',' +
-        this.a + ')';
-};
-
-// Color copying:
-
-Color.prototype.copy = function () {
-    return new Color(
-        this.r,
-        this.g,
-        this.b,
-        this.a
-    );
-};
-
-// Color comparison:
-
-Color.prototype.eq = function (aColor) {
-    // ==
-    return aColor &&
-        this.r === aColor.r &&
-        this.g === aColor.g &&
-        this.b === aColor.b;
-};
-
-// Color conversion (hsv):
-
-Color.prototype.hsv = function () {
-    // ignore alpha
-    var max, min, h, s, v, d,
-        rr = this.r / 255,
-        gg = this.g / 255,
-        bb = this.b / 255;
-    max = Math.max(rr, gg, bb);
-    min = Math.min(rr, gg, bb);
-    h = max;
-    s = max;
-    v = max;
-    d = max - min;
-    s = max === 0 ? 0 : d / max;
-    if (max === min) {
-        h = 0;
-    } else {
-        switch (max) {
-        case rr:
-            h = (gg - bb) / d + (gg < bb ? 6 : 0);
-            break;
-        case gg:
-            h = (bb - rr) / d + 2;
-            break;
-        case bb:
-            h = (rr - gg) / d + 4;
-            break;
-        }
-        h /= 6;
-    }
-    return [h, s, v];
-};
-
-Color.prototype.set_hsv = function (h, s, v) {
-    // ignore alpha, h, s and v are to be within [0, 1]
-    var i, f, p, q, t;
-    i = Math.floor(h * 6);
-    f = h * 6 - i;
-    p = v * (1 - s);
-    q = v * (1 - f * s);
-    t = v * (1 - (1 - f) * s);
-    switch (i % 6) {
-    case 0:
-        this.r = v;
-        this.g = t;
-        this.b = p;
-        break;
-    case 1:
-        this.r = q;
-        this.g = v;
-        this.b = p;
-        break;
-    case 2:
-        this.r = p;
-        this.g = v;
-        this.b = t;
-        break;
-    case 3:
-        this.r = p;
-        this.g = q;
-        this.b = v;
-        break;
-    case 4:
-        this.r = t;
-        this.g = p;
-        this.b = v;
-        break;
-    case 5:
-        this.r = v;
-        this.g = p;
-        this.b = q;
-        break;
-    }
-
-    this.r *= 255;
-    this.g *= 255;
-    this.b *= 255;
-
-};
-
-// Color mixing:
-
-Color.prototype.mixed = function (proportion, otherColor) {
-    // answer a copy of this color mixed with another color, ignore alpha
-    var frac1 = Math.min(Math.max(proportion, 0), 1),
-        frac2 = 1 - frac1;
-    return new Color(
-        this.r * frac1 + otherColor.r * frac2,
-        this.g * frac1 + otherColor.g * frac2,
-        this.b * frac1 + otherColor.b * frac2
-    );
-};
-
-Color.prototype.darker = function (percent) {
-    // return an rgb-interpolated darker copy of me, ignore alpha
-    var fract = 0.8333;
-    if (percent) {
-        fract = (100 - percent) / 100;
-    }
-    return this.mixed(fract, new Color(0, 0, 0));
-};
-
-Color.prototype.lighter = function (percent) {
-    // return an rgb-interpolated lighter copy of me, ignore alpha
-    var fract = 0.8333;
-    if (percent) {
-        fract = (100 - percent) / 100;
-    }
-    return this.mixed(fract, new Color(255, 255, 255));
-};
-
-Color.prototype.dansDarker = function () {
-    // return an hsv-interpolated darker copy of me, ignore alpha
-    var hsv = this.hsv(),
-        result = new Color(),
-        vv = Math.max(hsv[2] - 0.16, 0);
-    result.set_hsv(hsv[0], hsv[1], vv);
-    return result;
-};
-
-// Points //////////////////////////////////////////////////////////////
-
-// Point instance creation:
-
-function Point(x, y) {
-    this.x = x || 0;
-    this.y = y || 0;
-}
-
-// Point string representation: e.g. '12@68'
-
-Point.prototype.toString = function () {
-    return Math.round(this.x.toString()) +
-        '@' + Math.round(this.y.toString());
-};
-
-// Point copying:
-
-Point.prototype.copy = function () {
-    return new Point(this.x, this.y);
-};
-
-// Point comparison:
-
-Point.prototype.eq = function (aPoint) {
-    // ==
-    return this.x === aPoint.x && this.y === aPoint.y;
-};
-
-Point.prototype.lt = function (aPoint) {
-    // <
-    return this.x < aPoint.x && this.y < aPoint.y;
-};
-
-Point.prototype.gt = function (aPoint) {
-    // >
-    return this.x > aPoint.x && this.y > aPoint.y;
-};
-
-Point.prototype.ge = function (aPoint) {
-    // >=
-    return this.x >= aPoint.x && this.y >= aPoint.y;
-};
-
-Point.prototype.le = function (aPoint) {
-    // <=
-    return this.x <= aPoint.x && this.y <= aPoint.y;
-};
-
-Point.prototype.max = function (aPoint) {
-    return new Point(Math.max(this.x, aPoint.x),
-        Math.max(this.y, aPoint.y));
-};
-
-Point.prototype.min = function (aPoint) {
-    return new Point(Math.min(this.x, aPoint.x),
-        Math.min(this.y, aPoint.y));
-};
-
-// Point conversion:
-
-Point.prototype.round = function () {
-    return new Point(Math.round(this.x), Math.round(this.y));
-};
-
-Point.prototype.abs = function () {
-    return new Point(Math.abs(this.x), Math.abs(this.y));
-};
-
-Point.prototype.neg = function () {
-    return new Point(-this.x, -this.y);
-};
-
-Point.prototype.mirror = function () {
-    return new Point(this.y, this.x);
-};
-
-Point.prototype.floor = function () {
-    return new Point(
-        Math.max(Math.floor(this.x), 0),
-        Math.max(Math.floor(this.y), 0)
-    );
-};
-
-Point.prototype.ceil = function () {
-    return new Point(Math.ceil(this.x), Math.ceil(this.y));
-};
-
-// Point arithmetic:
-
-Point.prototype.add = function (other) {
-    if (other instanceof Point) {
-        return new Point(this.x + other.x, this.y + other.y);
-    }
-    return new Point(this.x + other, this.y + other);
-};
-
-Point.prototype.subtract = function (other) {
-    if (other instanceof Point) {
-        return new Point(this.x - other.x, this.y - other.y);
-    }
-    return new Point(this.x - other, this.y - other);
-};
-
-Point.prototype.multiplyBy = function (other) {
-    if (other instanceof Point) {
-        return new Point(this.x * other.x, this.y * other.y);
-    }
-    return new Point(this.x * other, this.y * other);
-};
-
-Point.prototype.divideBy = function (other) {
-    if (other instanceof Point) {
-        return new Point(this.x / other.x, this.y / other.y);
-    }
-    return new Point(this.x / other, this.y / other);
-};
-
-Point.prototype.floorDivideBy = function (other) {
-    if (other instanceof Point) {
-        return new Point(Math.floor(this.x / other.x),
-            Math.floor(this.y / other.y));
-    }
-    return new Point(Math.floor(this.x / other),
-        Math.floor(this.y / other));
-};
-
-// Point polar coordinates:
-
-Point.prototype.r = function () {
-    var t = (this.multiplyBy(this));
-    return Math.sqrt(t.x + t.y);
-};
-
-Point.prototype.degrees = function () {
-/*
-    answer the angle I make with origin in degrees.
-    Right is 0, down is 90
-*/
-    var tan, theta;
-
-    if (this.x === 0) {
-        if (this.y >= 0) {
-            return 90;
-        }
-        return 270;
-    }
-    tan = this.y / this.x;
-    theta = Math.atan(tan);
-    if (this.x >= 0) {
-        if (this.y >= 0) {
-            return degrees(theta);
-        }
-        return 360 + (degrees(theta));
-    }
-    return 180 + degrees(theta);
-};
-
-Point.prototype.theta = function () {
-/*
-    answer the angle I make with origin in radians.
-    Right is 0, down is 90
-*/
-    var tan, theta;
-
-    if (this.x === 0) {
-        if (this.y >= 0) {
-            return radians(90);
-        }
-        return radians(270);
-    }
-    tan = this.y / this.x;
-    theta = Math.atan(tan);
-    if (this.x >= 0) {
-        if (this.y >= 0) {
-            return theta;
-        }
-        return radians(360) + theta;
-    }
-    return radians(180) + theta;
-};
-
-// Point functions:
-
-Point.prototype.crossProduct = function (aPoint) {
-    return this.multiplyBy(aPoint.mirror());
-};
-
-Point.prototype.distanceTo = function (aPoint) {
-    return (aPoint.subtract(this)).r();
-};
-
-Point.prototype.rotate = function (direction, center) {
-    // direction must be 'right', 'left' or 'pi'
-    var offset = this.subtract(center);
-    if (direction === 'right') {
-        return new Point(-offset.y, offset.y).add(center);
-    }
-    if (direction === 'left') {
-        return new Point(offset.y, -offset.y).add(center);
-    }
-    // direction === 'pi'
-    return center.subtract(offset);
-};
-
-Point.prototype.flip = function (direction, center) {
-    // direction must be 'vertical' or 'horizontal'
-    if (direction === 'vertical') {
-        return new Point(this.x, center.y * 2 - this.y);
-    }
-    // direction === 'horizontal'
-    return new Point(center.x * 2 - this.x, this.y);
-};
-
-Point.prototype.distanceAngle = function (dist, angle) {
-    var deg = angle, x, y;
-    if (deg > 270) {
-        deg = deg - 360;
-    } else if (deg < -270) {
-        deg = deg + 360;
-    }
-    if (-90 <= deg && deg <= 90) {
-        x = Math.sin(radians(deg)) * dist;
-        y = Math.sqrt((dist * dist) - (x * x));
-        return new Point(x + this.x, this.y - y);
-    }
-    x = Math.sin(radians(180 - deg)) * dist;
-    y = Math.sqrt((dist * dist) - (x * x));
-    return new Point(x + this.x, this.y + y);
-};
-
-// Point transforming:
-
-Point.prototype.scaleBy = function (scalePoint) {
-    return this.multiplyBy(scalePoint);
-};
-
-Point.prototype.translateBy = function (deltaPoint) {
-    return this.add(deltaPoint);
-};
-
-Point.prototype.rotateBy = function (angle, centerPoint) {
-    var center = centerPoint || new Point(0, 0),
-        p = this.subtract(center),
-        r = p.r(),
-        theta = angle - p.theta();
-    return new Point(
-        center.x + (r * Math.cos(theta)),
-        center.y - (r * Math.sin(theta))
-    );
-};
-
-// Point conversion:
-
-Point.prototype.asArray = function () {
-    return [this.x, this.y];
-};
-
-// Rectangles //////////////////////////////////////////////////////////
-
-// Rectangle instance creation:
-
-function Rectangle(left, top, right, bottom) {
-    this.init(new Point((left || 0), (top || 0)),
-            new Point((right || 0), (bottom || 0)));
-}
-
-Rectangle.prototype.init = function (originPoint, cornerPoint) {
-    this.origin = originPoint;
-    this.corner = cornerPoint;
-};
-
-// Rectangle string representation: e.g. '[0@0 | 160@80]'
-
-Rectangle.prototype.toString = function () {
-    return '[' + this.origin.toString() + ' | ' +
-        this.extent().toString() + ']';
-};
-
-// Rectangle copying:
-
-Rectangle.prototype.copy = function () {
-    return new Rectangle(
-        this.left(),
-        this.top(),
-        this.right(),
-        this.bottom()
-    );
-};
-
-// creating Rectangle instances from Points:
-
-Point.prototype.corner = function (cornerPoint) {
-    // answer a new Rectangle
-    return new Rectangle(
-        this.x,
-        this.y,
-        cornerPoint.x,
-        cornerPoint.y
-    );
-};
-
-Point.prototype.rectangle = function (aPoint) {
-    // answer a new Rectangle
-    var org, crn;
-    org = this.min(aPoint);
-    crn = this.max(aPoint);
-    return new Rectangle(org.x, org.y, crn.x, crn.y);
-};
-
-Point.prototype.extent = function (aPoint) {
-    //answer a new Rectangle
-    var crn = this.add(aPoint);
-    return new Rectangle(this.x, this.y, crn.x, crn.y);
-};
-
-// Rectangle accessing - setting:
-
-Rectangle.prototype.setTo = function (left, top, right, bottom) {
-    // note: all inputs are optional and can be omitted
-
-    this.origin = new Point(
-        left || ((left === 0) ? 0 : this.left()),
-        top || ((top === 0) ? 0 : this.top())
-    );
-
-    this.corner = new Point(
-        right || ((right === 0) ? 0 : this.right()),
-        bottom || ((bottom === 0) ? 0 : this.bottom())
-    );
-};
-
-// Rectangle accessing - getting:
-
-Rectangle.prototype.area = function () {
-    //requires width() and height() to be defined
-    var w = this.width();
-    if (w < 0) {
-        return 0;
-    }
-    return Math.max(w * this.height(), 0);
-};
-
-Rectangle.prototype.bottom = function () {
-    return this.corner.y;
-};
-
-Rectangle.prototype.bottomCenter = function () {
-    return new Point(this.center().x, this.bottom());
-};
-
-Rectangle.prototype.bottomLeft = function () {
-    return new Point(this.origin.x, this.corner.y);
-};
-
-Rectangle.prototype.bottomRight = function () {
-    return this.corner.copy();
-};
-
-Rectangle.prototype.boundingBox = function () {
-    return this;
-};
-
-Rectangle.prototype.center = function () {
-    return this.origin.add(
-        this.corner.subtract(this.origin).floorDivideBy(2)
-    );
-};
-
-Rectangle.prototype.corners = function () {
-    return [this.origin,
-        this.bottomLeft(),
-        this.corner,
-        this.topRight()];
-};
-
-Rectangle.prototype.extent = function () {
-    return this.corner.subtract(this.origin);
-};
-
-Rectangle.prototype.height = function () {
-    return this.corner.y - this.origin.y;
-};
-
-Rectangle.prototype.left = function () {
-    return this.origin.x;
-};
-
-Rectangle.prototype.leftCenter = function () {
-    return new Point(this.left(), this.center().y);
-};
-
-Rectangle.prototype.right = function () {
-    return this.corner.x;
-};
-
-Rectangle.prototype.rightCenter = function () {
-    return new Point(this.right(), this.center().y);
-};
-
-Rectangle.prototype.top = function () {
-    return this.origin.y;
-};
-
-Rectangle.prototype.topCenter = function () {
-    return new Point(this.center().x, this.top());
-};
-
-Rectangle.prototype.topLeft = function () {
-    return this.origin;
-};
-
-Rectangle.prototype.topRight = function () {
-    return new Point(this.corner.x, this.origin.y);
-};
-
-Rectangle.prototype.width = function () {
-    return this.corner.x - this.origin.x;
-};
-
-Rectangle.prototype.position = function () {
-    return this.origin;
-};
-
-// Rectangle comparison:
-
-Rectangle.prototype.eq = function (aRect) {
-    return this.origin.eq(aRect.origin) &&
-        this.corner.eq(aRect.corner);
-};
-
-Rectangle.prototype.abs = function () {
-    var newOrigin, newCorner;
-
-    newOrigin = this.origin.abs();
-    newCorner = this.corner.max(newOrigin);
-    return newOrigin.corner(newCorner);
-};
-
-// Rectangle functions:
-
-Rectangle.prototype.insetBy = function (delta) {
-    // delta can be either a Point or a Number
-    var result = new Rectangle();
-    result.origin = this.origin.add(delta);
-    result.corner = this.corner.subtract(delta);
-    return result;
-};
-
-Rectangle.prototype.expandBy = function (delta) {
-    // delta can be either a Point or a Number
-    var result = new Rectangle();
-    result.origin = this.origin.subtract(delta);
-    result.corner = this.corner.add(delta);
-    return result;
-};
-
-Rectangle.prototype.growBy = function (delta) {
-    // delta can be either a Point or a Number
-    var result = new Rectangle();
-    result.origin = this.origin.copy();
-    result.corner = this.corner.add(delta);
-    return result;
-};
-
-Rectangle.prototype.intersect = function (aRect) {
-    var result = new Rectangle();
-    result.origin = this.origin.max(aRect.origin);
-    result.corner = this.corner.min(aRect.corner);
-    return result;
-};
-
-Rectangle.prototype.merge = function (aRect) {
-    var result = new Rectangle();
-    result.origin = this.origin.min(aRect.origin);
-    result.corner = this.corner.max(aRect.corner);
-    return result;
-};
-
-Rectangle.prototype.mergeWith = function (aRect) {
-    // mutates myself
-    this.origin = this.origin.min(aRect.origin);
-    this.corner = this.corner.max(aRect.corner);
-};
-
-Rectangle.prototype.round = function () {
-    return this.origin.round().corner(this.corner.round());
-};
-
-Rectangle.prototype.spread = function () {
-    // round me by applying floor() to my origin and ceil() to my corner
-    return this.origin.floor().corner(this.corner.ceil());
-};
-
-Rectangle.prototype.amountToTranslateWithin = function (aRect) {
-/*
-    Answer a Point, delta, such that self + delta is forced within
-    aRectangle. when all of me cannot be made to fit, prefer to keep
-    my topLeft inside. Taken from Squeak.
-*/
-    var dx = 0, dy = 0;
-
-    if (this.right() > aRect.right()) {
-        dx = aRect.right() - this.right();
-    }
-    if (this.bottom() > aRect.bottom()) {
-        dy = aRect.bottom() - this.bottom();
-    }
-    if ((this.left() + dx) < aRect.left()) {
-        dx = aRect.left() - this.left();
-    }
-    if ((this.top() + dy) < aRect.top()) {
-        dy = aRect.top() - this.top();
-    }
-    return new Point(dx, dy);
-};
-
-// Rectangle testing:
-
-Rectangle.prototype.containsPoint = function (aPoint) {
-    return this.origin.le(aPoint) && aPoint.lt(this.corner);
-};
-
-Rectangle.prototype.containsRectangle = function (aRect) {
-    return aRect.origin.gt(this.origin) &&
-        aRect.corner.lt(this.corner);
-};
-
-Rectangle.prototype.intersects = function (aRect) {
-    var ro = aRect.origin, rc = aRect.corner;
-    return (rc.x >= this.origin.x) &&
-        (rc.y >= this.origin.y) &&
-        (ro.x <= this.corner.x) &&
-        (ro.y <= this.corner.y);
-};
-
-Rectangle.prototype.isNearTo = function (aRect, threshold) {
-    var ro = aRect.origin, rc = aRect.corner, border = threshold || 0;
-    return (rc.x + border >= this.origin.x) &&
-        (rc.y  + border >= this.origin.y) &&
-        (ro.x - border <= this.corner.x) &&
-        (ro.y - border <= this.corner.y);
-};
-
-// Rectangle transforming:
-
-Rectangle.prototype.scaleBy = function (scale) {
-    // scale can be either a Point or a scalar
-    var o = this.origin.multiplyBy(scale),
-        c = this.corner.multiplyBy(scale);
-    return new Rectangle(o.x, o.y, c.x, c.y);
-};
-
-Rectangle.prototype.translateBy = function (factor) {
-    // factor can be either a Point or a scalar
-    var o = this.origin.add(factor),
-        c = this.corner.add(factor);
-    return new Rectangle(o.x, o.y, c.x, c.y);
-};
-
-// Rectangle converting:
-
-Rectangle.prototype.asArray = function () {
-    return [this.left(), this.top(), this.right(), this.bottom()];
-};
-
-Rectangle.prototype.asArray_xywh = function () {
-    return [this.left(), this.top(), this.width(), this.height()];
-};
-
-// Nodes ///////////////////////////////////////////////////////////////
-
-// Node instance creation:
-
-function Node(parent, childrenArray) {
-    this.init(parent || null, childrenArray || []);
-}
-
-Node.prototype.init = function (parent, childrenArray) {
-    this.parent = parent || null;
-    this.children = childrenArray || [];
-};
-
-// Node string representation: e.g. 'a Node[3]'
-
-Node.prototype.toString = function () {
-    return 'a Node' + '[' + this.children.length.toString() + ']';
-};
-
-// Node accessing:
-
-Node.prototype.addChild = function (aNode) {
-    this.children.push(aNode);
-    aNode.parent = this;
-};
-
-Node.prototype.addChildFirst = function (aNode) {
-    this.children.splice(0, null, aNode);
-    aNode.parent = this;
-};
-
-Node.prototype.removeChild = function (aNode) {
-    var idx = this.children.indexOf(aNode);
-    if (idx !== -1) {
-        this.children.splice(idx, 1);
-    }
-};
-
-// Node functions:
-
-Node.prototype.root = function () {
-    if (this.parent === null) {
-        return this;
-    }
-    return this.parent.root();
-};
-
-Node.prototype.depth = function () {
-    if (this.parent === null) {
-        return 0;
-    }
-    return this.parent.depth() + 1;
-};
-
-Node.prototype.allChildren = function () {
-    // includes myself
-    var result = [this];
-    this.children.forEach(function (child) {
-        result = result.concat(child.allChildren());
-    });
-    return result;
-};
-
-Node.prototype.forAllChildren = function (aFunction) {
-    if (this.children.length > 0) {
-        this.children.forEach(function (child) {
-            child.forAllChildren(aFunction);
-        });
-    }
-    aFunction.call(null, this);
-};
-
-Node.prototype.allLeafs = function () {
-    var result = [];
-    this.allChildren().forEach(function (element) {
-        if (element.children.length === 0) {
-            result.push(element);
-        }
-    });
-    return result;
-};
-
-Node.prototype.allParents = function () {
-    // includes myself
-    var result = [this];
-    if (this.parent !== null) {
-        result = result.concat(this.parent.allParents());
-    }
-    return result;
-};
-
-Node.prototype.siblings = function () {
-    var myself = this;
-    if (this.parent === null) {
-        return [];
-    }
-    return this.parent.children.filter(function (child) {
-        return child !== myself;
-    });
-};
-
-Node.prototype.parentThatIsA = function (constructor) {
-    // including myself
-    if (this instanceof constructor) {
-        return this;
-    }
-    if (!this.parent) {
-        return null;
-    }
-    return this.parent.parentThatIsA(constructor);
-};
-
-Node.prototype.parentThatIsAnyOf = function (constructors) {
-    // including myself
-    var yup = false,
-        myself = this;
-    constructors.forEach(function (each) {
-        if (myself.constructor === each) {
-            yup = true;
-            return;
-        }
-    });
-    if (yup) {
-        return this;
-    }
-    if (!this.parent) {
-        return null;
-    }
-    return this.parent.parentThatIsAnyOf(constructors);
-};
-
-// Morphs //////////////////////////////////////////////////////////////
-
-// Morph: referenced constructors
-
-var Morph;
-var WorldMorph;
-var HandMorph;
-var ShadowMorph;
-var FrameMorph;
-var MenuMorph;
-var HandleMorph;
-var StringFieldMorph;
-var ColorPickerMorph;
-var SliderMorph;
-var ScrollFrameMorph;
-var InspectorMorph;
-var StringMorph;
-var TextMorph;
-
-// Morph inherits from Node:
-
-Morph.prototype = new Node();
-Morph.prototype.constructor = Morph;
-Morph.uber = Node.prototype;
-
-// Morph settings:
-
-/*
-    damage list housekeeping
-
-    the trackChanges property of the Morph prototype is a Boolean switch
-    that determines whether the World's damage list ('broken' rectangles)
-    tracks changes. By default the switch is always on. If set to false
-    changes are not stored. This can be very useful for housekeeping of
-    the damage list in situations where a large number of (sub-) morphs
-    are changed more or less at once. Instead of keeping track of every
-    single submorph's changes tremendous performance improvements can be
-    achieved by setting the trackChanges flag to false before propagating
-    the layout changes, setting it to true again and then storing the full
-    bounds of the surrounding morph. An an example refer to the
-
-        fixLayout()
-
-    method of InspectorMorph, or the
-
-        startLayout()
-        endLayout()
-
-    methods of SyntaxElementMorph in the Snap application.
-*/
-
-Morph.prototype.trackChanges = true;
-Morph.prototype.shadowBlur = 4;
-
-// Morph instance creation:
-
-function Morph() {
-    this.init();
-}
-
-// Morph initialization:
-
-Morph.prototype.init = function () {
-    Morph.uber.init.call(this);
-    this.isMorph = true;
-    this.bounds = new Rectangle(0, 0, 50, 40);
-    this.color = new Color(80, 80, 80);
-    this.texture = null; // optional url of a fill-image
-    this.cachedTexture = null; // internal cache of actual bg image
-    this.alpha = 1;
-    this.isVisible = true;
-    this.isDraggable = false;
-    this.isTemplate = false;
-    this.acceptsDrops = false;
-    this.noticesTransparentClick = false;
-    this.drawNew();
-    this.fps = 0;
-    this.customContextMenu = null;
-    this.lastTime = Date.now();
-    this.onNextStep = null; // optional function to be run once
-};
-
-// Morph string representation: e.g. 'a Morph 2 [20@45 | 130@250]'
-
-Morph.prototype.toString = function () {
-    return 'a ' +
-        (this.constructor.name ||
-            this.constructor.toString().split(' ')[1].split('(')[0]) +
-        ' ' +
-        this.children.length.toString() + ' ' +
-        this.bounds;
-};
-
-// Morph deleting:
-
-Morph.prototype.destroy = function () {
-    if (this.parent !== null) {
-        this.fullChanged();
-        this.parent.removeChild(this);
-    }
-};
-
-// Morph stepping:
-
-Morph.prototype.stepFrame = function () {
-    if (!this.step) {
-        return null;
-    }
-    var current, elapsed, leftover, nxt;
-    current = Date.now();
-    elapsed = current - this.lastTime;
-    if (this.fps > 0) {
-        leftover = (1000 / this.fps) - elapsed;
-    } else {
-        leftover = 0;
-    }
-    if (leftover < 1) {
-        this.lastTime = current;
-        if (this.onNextStep) {
-            nxt = this.onNextStep;
-            this.onNextStep = null;
-            nxt.call(this);
-        }
-        this.step();
-        this.children.forEach(function (child) {
-            child.stepFrame();
-        });
-    }
-};
-
-Morph.prototype.nextSteps = function (arrayOfFunctions) {
-    var lst = arrayOfFunctions || [],
-        nxt = lst.shift(),
-        myself = this;
-    if (nxt) {
-        this.onNextStep = function () {
-            nxt.call(myself);
-            myself.nextSteps(lst);
-        };
-    }
-};
-
-Morph.prototype.step = function () {
-    nop();
-};
-
-// Morph accessing - geometry getting:
-
-Morph.prototype.left = function () {
-    return this.bounds.left();
-};
-
-Morph.prototype.right = function () {
-    return this.bounds.right();
-};
-
-Morph.prototype.top = function () {
-    return this.bounds.top();
-};
-
-Morph.prototype.bottom = function () {
-    return this.bounds.bottom();
-};
-
-Morph.prototype.center = function () {
-    return this.bounds.center();
-};
-
-Morph.prototype.bottomCenter = function () {
-    return this.bounds.bottomCenter();
-};
-
-Morph.prototype.bottomLeft = function () {
-    return this.bounds.bottomLeft();
-};
-
-Morph.prototype.bottomRight = function () {
-    return this.bounds.bottomRight();
-};
-
-Morph.prototype.boundingBox = function () {
-    return this.bounds;
-};
-
-Morph.prototype.corners = function () {
-    return this.bounds.corners();
-};
-
-Morph.prototype.leftCenter = function () {
-    return this.bounds.leftCenter();
-};
-
-Morph.prototype.rightCenter = function () {
-    return this.bounds.rightCenter();
-};
-
-Morph.prototype.topCenter = function () {
-    return this.bounds.topCenter();
-};
-
-Morph.prototype.topLeft = function () {
-    return this.bounds.topLeft();
-};
-
-Morph.prototype.topRight = function () {
-    return this.bounds.topRight();
-};
-Morph.prototype.position = function () {
-    return this.bounds.origin;
-};
-
-Morph.prototype.extent = function () {
-    return this.bounds.extent();
-};
-
-Morph.prototype.width = function () {
-    return this.bounds.width();
-};
-
-Morph.prototype.height = function () {
-    return this.bounds.height();
-};
-
-Morph.prototype.fullBounds = function () {
-    var result;
-    result = this.bounds;
-    this.children.forEach(function (child) {
-        if (child.isVisible) {
-            result = result.merge(child.fullBounds());
-        }
-    });
-    return result;
-};
-
-Morph.prototype.fullBoundsNoShadow = function () {
-    // answer my full bounds but ignore any shadow
-    var result;
-    result = this.bounds;
-    this.children.forEach(function (child) {
-        if (!(child instanceof ShadowMorph) && (child.isVisible)) {
-            result = result.merge(child.fullBounds());
-        }
-    });
-    return result;
-};
-
-Morph.prototype.visibleBounds = function () {
-    // answer which part of me is not clipped by a Frame
-    var visible = this.bounds,
-        frames = this.allParents().filter(function (p) {
-            return p instanceof FrameMorph;
-        });
-    frames.forEach(function (f) {
-        visible = visible.intersect(f.bounds);
-    });
-    return visible;
-};
-
-// Morph accessing - simple changes:
-
-Morph.prototype.moveBy = function (delta) {
-    this.changed();
-    this.bounds = this.bounds.translateBy(delta);
-    this.children.forEach(function (child) {
-        child.moveBy(delta);
-    });
-    this.changed();
-};
-
-Morph.prototype.silentMoveBy = function (delta) {
-    this.bounds = this.bounds.translateBy(delta);
-    this.children.forEach(function (child) {
-        child.silentMoveBy(delta);
-    });
-};
-
-Morph.prototype.setPosition = function (aPoint) {
-    var delta = aPoint.subtract(this.topLeft());
-    if ((delta.x !== 0) || (delta.y !== 0)) {
-        this.moveBy(delta);
-    }
-};
-
-Morph.prototype.silentSetPosition = function (aPoint) {
-    var delta = aPoint.subtract(this.topLeft());
-    if ((delta.x !== 0) || (delta.y !== 0)) {
-        this.silentMoveBy(delta);
-    }
-};
-
-Morph.prototype.setLeft = function (x) {
-    this.setPosition(
-        new Point(
-            x,
-            this.top()
-        )
-    );
-};
-
-Morph.prototype.setRight = function (x) {
-    this.setPosition(
-        new Point(
-            x - this.width(),
-            this.top()
-        )
-    );
-};
-
-Morph.prototype.setTop = function (y) {
-    this.setPosition(
-        new Point(
-            this.left(),
-            y
-        )
-    );
-};
-
-Morph.prototype.setBottom = function (y) {
-    this.setPosition(
-        new Point(
-            this.left(),
-            y - this.height()
-        )
-    );
-};
-
-Morph.prototype.setCenter = function (aPoint) {
-    this.setPosition(
-        aPoint.subtract(
-            this.extent().floorDivideBy(2)
-        )
-    );
-};
-
-Morph.prototype.setFullCenter = function (aPoint) {
-    this.setPosition(
-        aPoint.subtract(
-            this.fullBounds().extent().floorDivideBy(2)
-        )
-    );
-};
-
-Morph.prototype.keepWithin = function (aMorph) {
-    // make sure I am completely within another Morph's bounds
-    var leftOff, rightOff, topOff, bottomOff;
-    leftOff = this.fullBounds().left() - aMorph.left();
-    if (leftOff < 0) {
-        this.moveBy(new Point(-leftOff, 0));
-    }
-    rightOff = this.fullBounds().right() - aMorph.right();
-    if (rightOff > 0) {
-        this.moveBy(new Point(-rightOff, 0));
-    }
-    topOff = this.fullBounds().top() - aMorph.top();
-    if (topOff < 0) {
-        this.moveBy(new Point(0, -topOff));
-    }
-    bottomOff = this.fullBounds().bottom() - aMorph.bottom();
-    if (bottomOff > 0) {
-        this.moveBy(new Point(0, -bottomOff));
-    }
-};
-
-// Morph accessing - dimensional changes requiring a complete redraw
-
-Morph.prototype.setExtent = function (aPoint) {
-    if (!aPoint.eq(this.extent())) {
-        this.changed();
-        this.silentSetExtent(aPoint);
-        this.changed();
-        this.drawNew();
-    }
-};
-
-Morph.prototype.silentSetExtent = function (aPoint) {
-    var ext, newWidth, newHeight;
-    ext = aPoint.round();
-    newWidth = Math.max(ext.x, 0);
-    newHeight = Math.max(ext.y, 0);
-    this.bounds.corner = new Point(
-        this.bounds.origin.x + newWidth,
-        this.bounds.origin.y + newHeight
-    );
-};
-
-Morph.prototype.setWidth = function (width) {
-    this.setExtent(new Point(width || 0, this.height()));
-};
-
-Morph.prototype.silentSetWidth = function (width) {
-    // do not drawNew() just yet
-    var w = Math.max(Math.round(width || 0), 0);
-    this.bounds.corner = new Point(
-        this.bounds.origin.x + w,
-        this.bounds.corner.y
-    );
-};
-
-Morph.prototype.setHeight = function (height) {
-    this.setExtent(new Point(this.width(), height || 0));
-};
-
-Morph.prototype.silentSetHeight = function (height) {
-    // do not drawNew() just yet
-    var h = Math.max(Math.round(height || 0), 0);
-    this.bounds.corner = new Point(
-        this.bounds.corner.x,
-        this.bounds.origin.y + h
-    );
-};
-
-Morph.prototype.setColor = function (aColor) {
-    if (aColor) {
-        if (!this.color.eq(aColor)) {
-            this.color = aColor;
-            this.changed();
-            this.drawNew();
-        }
-    }
-};
-
-// Morph displaying:
-
-Morph.prototype.drawNew = function () {
-    // initialize my surface property
-    this.image = newCanvas(this.extent());
-    var context = this.image.getContext('2d');
-    context.fillStyle = this.color.toString();
-    context.fillRect(0, 0, this.width(), this.height());
-    if (this.cachedTexture) {
-        this.drawCachedTexture();
-    } else if (this.texture) {
-        this.drawTexture(this.texture);
-    }
-};
-
-Morph.prototype.drawTexture = function (url) {
-    var myself = this;
-    this.cachedTexture = new Image();
-    this.cachedTexture.onload = function () {
-        myself.drawCachedTexture();
-    };
-    this.cachedTexture.src = this.texture = url; // make absolute
-};
-
-Morph.prototype.drawCachedTexture = function () {
-    var bg = this.cachedTexture,
-        cols = Math.floor(this.image.width / bg.width),
-        lines = Math.floor(this.image.height / bg.height),
-        x,
-        y,
-        context = this.image.getContext('2d');
-
-    for (y = 0; y <= lines; y += 1) {
-        for (x = 0; x <= cols; x += 1) {
-            context.drawImage(bg, x * bg.width, y * bg.height);
-        }
-    }
-    this.changed();
-};
-
-/*
-Morph.prototype.drawCachedTexture = function () {
-    var context = this.image.getContext('2d'),
-        pattern = context.createPattern(this.cachedTexture, 'repeat');
-    context.fillStyle = pattern;
-    context.fillRect(0, 0, this.image.width, this.image.height);
-    this.changed();
-};
-*/
-
-Morph.prototype.drawOn = function (aCanvas, aRect) {
-    var rectangle, area, delta, src, context, w, h, sl, st;
-    if (!this.isVisible) {
-        return null;
-    }
-    rectangle = aRect || this.bounds();
-    area = rectangle.intersect(this.bounds).round();
-    if (area.extent().gt(new Point(0, 0))) {
-        delta = this.position().neg();
-        src = area.copy().translateBy(delta).round();
-        context = aCanvas.getContext('2d');
-        context.globalAlpha = this.alpha;
-
-        sl = src.left();
-        st = src.top();
-        w = Math.min(src.width(), this.image.width - sl);
-        h = Math.min(src.height(), this.image.height - st);
-
-        if (w < 1 || h < 1) {
-            return null;
-        }
-
-        context.drawImage(
-            this.image,
-            src.left(),
-            src.top(),
-            w,
-            h,
-            area.left(),
-            area.top(),
-            w,
-            h
-        );
-
-    /* "for debugging purposes:"
-
-        try {
-            context.drawImage(
-                this.image,
-                src.left(),
-                src.top(),
-                w,
-                h,
-                area.left(),
-                area.top(),
-                w,
-                h
-            );
-        } catch (err) {
-            alert('internal error\n\n' + err
-                + '\n ---'
-                + '\n canvas: ' + aCanvas
-                + '\n canvas.width: ' + aCanvas.width
-                + '\n canvas.height: ' + aCanvas.height
-                + '\n ---'
-                + '\n image: ' + this.image
-                + '\n image.width: ' + this.image.width
-                + '\n image.height: ' + this.image.height
-                + '\n ---'
-                + '\n w: ' + w
-                + '\n h: ' + h
-                + '\n sl: ' + sl
-                + '\n st: ' + st
-                + '\n area.left: ' + area.left()
-                + '\n area.top ' + area.top()
-                );
-        }
-    */
-
-    }
-};
-
-Morph.prototype.fullDrawOn = function (aCanvas, aRect) {
-    var rectangle;
-    if (!this.isVisible) {
-        return null;
-    }
-    rectangle = aRect || this.fullBounds();
-    this.drawOn(aCanvas, rectangle);
-    this.children.forEach(function (child) {
-        child.fullDrawOn(aCanvas, rectangle);
-    });
-};
-
-Morph.prototype.hide = function () {
-    this.isVisible = false;
-    this.changed();
-    this.children.forEach(function (child) {
-        child.hide();
-    });
-};
-
-Morph.prototype.show = function () {
-    this.isVisible = true;
-    this.changed();
-    this.children.forEach(function (child) {
-        child.show();
-    });
-};
-
-Morph.prototype.toggleVisibility = function () {
-    this.isVisible = (!this.isVisible);
-    this.changed();
-    this.children.forEach(function (child) {
-        child.toggleVisibility();
-    });
-};
-
-// Morph full image:
-
-Morph.prototype.fullImageClassic = function () {
-    // why doesn't this work for all Morphs?
-    var fb = this.fullBounds(),
-        img = newCanvas(fb.extent()),
-        ctx = img.getContext('2d');
-    ctx.translate(-this.bounds.origin.x, -this.bounds.origin.y);
-    this.fullDrawOn(img, fb);
-    img.globalAlpha = this.alpha;
-    return img;
-};
-
-Morph.prototype.fullImage = function () {
-    var img, ctx, fb;
-    img = newCanvas(this.fullBounds().extent());
-    ctx = img.getContext('2d');
-    fb = this.fullBounds();
-    this.allChildren().forEach(function (morph) {
-        if (morph.isVisible) {
-            ctx.globalAlpha = morph.alpha;
-            if (morph.image.width && morph.image.height) {
-                ctx.drawImage(
-                    morph.image,
-                    morph.bounds.origin.x - fb.origin.x,
-                    morph.bounds.origin.y - fb.origin.y
-                );
-            }
-        }
-    });
-    return img;
-};
-
-// Morph shadow:
-
-Morph.prototype.shadowImage = function (off, color) {
-    // fallback for Windows Chrome-Shadow bug
-    var fb, img, outline, sha, ctx,
-        offset = off || new Point(7, 7),
-        clr = color || new Color(0, 0, 0);
-    fb = this.fullBounds().extent();
-    img = this.fullImage();
-    outline = newCanvas(fb);
-    ctx = outline.getContext('2d');
-    ctx.drawImage(img, 0, 0);
-    ctx.globalCompositeOperation = 'destination-out';
-    ctx.drawImage(
-        img,
-        -offset.x,
-        -offset.y
-    );
-    sha = newCanvas(fb);
-    ctx = sha.getContext('2d');
-    ctx.drawImage(outline, 0, 0);
-    ctx.globalCompositeOperation = 'source-atop';
-    ctx.fillStyle = clr.toString();
-    ctx.fillRect(0, 0, fb.x, fb.y);
-    return sha;
-};
-
-Morph.prototype.shadowImageBlurred = function (off, color) {
-    var fb, img, sha, ctx,
-        offset = off || new Point(7, 7),
-        blur = this.shadowBlur,
-        clr = color || new Color(0, 0, 0);
-    fb = this.fullBounds().extent().add(blur * 2);
-    img = this.fullImage();
-    sha = newCanvas(fb);
-    ctx = sha.getContext('2d');
-    ctx.shadowOffsetX = offset.x;
-    ctx.shadowOffsetY = offset.y;
-    ctx.shadowBlur = blur;
-    ctx.shadowColor = clr.toString();
-    ctx.drawImage(
-        img,
-        blur - offset.x,
-        blur - offset.y
-    );
-    ctx.shadowOffsetX = 0;
-    ctx.shadowOffsetY = 0;
-    ctx.shadowBlur = 0;
-    ctx.globalCompositeOperation = 'destination-out';
-    ctx.drawImage(
-        img,
-        blur - offset.x,
-        blur - offset.y
-    );
-    return sha;
-};
-
-Morph.prototype.shadow = function (off, a, color) {
-    var shadow = new ShadowMorph(),
-        offset = off || new Point(7, 7),
-        alpha = a || ((a === 0) ? 0 : 0.2),
-        fb = this.fullBounds();
-    shadow.setExtent(fb.extent().add(this.shadowBlur * 2));
-    if (useBlurredShadows && !MorphicPreferences.isFlat) {
-        shadow.image = this.shadowImageBlurred(offset, color);
-        shadow.alpha = alpha;
-        shadow.setPosition(fb.origin.add(offset).subtract(this.shadowBlur));
-    } else {
-        shadow.image = this.shadowImage(offset, color);
-        shadow.alpha = alpha;
-        shadow.setPosition(fb.origin.add(offset));
-    }
-    return shadow;
-};
-
-Morph.prototype.addShadow = function (off, a, color) {
-    var shadow,
-        offset = off || new Point(7, 7),
-        alpha = a || ((a === 0) ? 0 : 0.2);
-    shadow = this.shadow(offset, alpha, color);
-    this.addBack(shadow);
-    this.fullChanged();
-    return shadow;
-};
-
-Morph.prototype.getShadow = function () {
-    var shadows;
-    shadows = this.children.slice(0).reverse().filter(
-        function (child) {
-            return child instanceof ShadowMorph;
-        }
-    );
-    if (shadows.length !== 0) {
-        return shadows[0];
-    }
-    return null;
-};
-
-Morph.prototype.removeShadow = function () {
-    var shadow = this.getShadow();
-    if (shadow !== null) {
-        this.fullChanged();
-        this.removeChild(shadow);
-    }
-};
-
-// Morph pen trails:
-
-Morph.prototype.penTrails = function () {
-    // answer my pen trails canvas. default is to answer my image
-    return this.image;
-};
-
-// Morph updating:
-
-Morph.prototype.changed = function () {
-    if (this.trackChanges) {
-        var w = this.root();
-        if (w instanceof WorldMorph) {
-            w.broken.push(this.visibleBounds().spread());
-        }
-    }
-    if (this.parent) {
-        this.parent.childChanged(this);
-    }
-};
-
-Morph.prototype.fullChanged = function () {
-    if (this.trackChanges) {
-        var w = this.root();
-        if (w instanceof WorldMorph) {
-            w.broken.push(this.fullBounds().spread());
-        }
-    }
-};
-
-Morph.prototype.childChanged = function () {
-    // react to a  change in one of my children,
-    // default is to just pass this message on upwards
-    // override this method for Morphs that need to adjust accordingly
-    if (this.parent) {
-        this.parent.childChanged(this);
-    }
-};
-
-// Morph accessing - structure:
-
-Morph.prototype.world = function () {
-    var root = this.root();
-    if (root instanceof WorldMorph) {
-        return root;
-    }
-    if (root instanceof HandMorph) {
-        return root.world;
-    }
-    return null;
-};
-
-Morph.prototype.add = function (aMorph) {
-    var owner = aMorph.parent;
-    if (owner !== null) {
-        owner.removeChild(aMorph);
-    }
-    this.addChild(aMorph);
-};
-
-Morph.prototype.addBack = function (aMorph) {
-    var owner = aMorph.parent;
-    if (owner !== null) {
-        owner.removeChild(aMorph);
-    }
-    this.addChildFirst(aMorph);
-};
-
-Morph.prototype.topMorphSuchThat = function (predicate) {
-    var next;
-    if (predicate.call(null, this)) {
-        next = detect(
-            this.children.slice(0).reverse(),
-            predicate
-        );
-        if (next) {
-            return next.topMorphSuchThat(predicate);
-        }
-        return this;
-    }
-    return null;
-};
-
-Morph.prototype.morphAt = function (aPoint) {
-    var morphs = this.allChildren().slice(0).reverse(),
-        result = null;
-    morphs.forEach(function (m) {
-        if (m.fullBounds().containsPoint(aPoint) &&
-                (result === null)) {
-            result = m;
-        }
-    });
-    return result;
-};
-
-/*
-    alternative -  more elegant and possibly more
-    performant - solution for morphAt.
-    Has some issues, commented out for now
-
-Morph.prototype.morphAt = function (aPoint) {
-    return this.topMorphSuchThat(function (m) {
-        return m.fullBounds().containsPoint(aPoint);
-    });
-};
-*/
-
-Morph.prototype.overlappedMorphs = function () {
-    //exclude the World
-    var world = this.world(),
-        fb = this.fullBounds(),
-        myself = this,
-        allParents = this.allParents(),
-        allChildren = this.allChildren(),
-        morphs;
-
-    morphs = world.allChildren();
-    return morphs.filter(function (m) {
-        return m.isVisible &&
-            m !== myself &&
-            m !== world &&
-            !contains(allParents, m) &&
-            !contains(allChildren, m) &&
-            m.fullBounds().intersects(fb);
-    });
-};
-
-// Morph pixel access:
-
-Morph.prototype.getPixelColor = function (aPoint) {
-    var point, context, data;
-    point = aPoint.subtract(this.bounds.origin);
-    context = this.image.getContext('2d');
-    data = context.getImageData(point.x, point.y, 1, 1);
-    return new Color(
-        data.data[0],
-        data.data[1],
-        data.data[2],
-        data.data[3]
-    );
-};
-
-Morph.prototype.isTransparentAt = function (aPoint) {
-    var point, context, data;
-    if (this.bounds.containsPoint(aPoint)) {
-        if (this.texture) {
-            return false;
-        }
-        point = aPoint.subtract(this.bounds.origin);
-        context = this.image.getContext('2d');
-        data = context.getImageData(
-            Math.floor(point.x),
-            Math.floor(point.y),
-            1,
-            1
-        );
-        return data.data[3] === 0;
-    }
-    return false;
-};
-
-// Morph duplicating:
-
-Morph.prototype.copy = function () {
-    var c = copy(this);
-    c.parent = null;
-    c.children = [];
-    c.bounds = this.bounds.copy();
-    return c;
-};
-
-Morph.prototype.fullCopy = function () {
-    /*
-    Produce a copy of me with my entire tree of submorphs. Morphs
-    mentioned more than once are all directed to a single new copy.
-    Other properties are also *shallow* copied, so you must override
-    to deep copy Arrays and (complex) Objects
-    */
-    var dict = {}, c;
-    c = this.copyRecordingReferences(dict);
-    c.forAllChildren(function (m) {
-        m.updateReferences(dict);
-    });
-    return c;
-};
-
-Morph.prototype.copyRecordingReferences = function (dict) {
-    /*
-    Recursively copy this entire composite morph, recording the
-    correspondence between old and new morphs in the given dictionary.
-    This dictionary will be used to update intra-composite references
-    in the copy. See updateReferences().
-    Note: This default implementation copies ONLY morphs in the
-    submorph hierarchy. If a morph stores morphs in other properties
-    that it wants to copy, then it should override this method to do so.
-    The same goes for morphs that contain other complex data that
-    should be copied when the morph is duplicated.
-    */
-    var c = this.copy();
-    dict[this] = c;
-    this.children.forEach(function (m) {
-        c.add(m.copyRecordingReferences(dict));
-    });
-    return c;
-};
-
-Morph.prototype.updateReferences = function (dict) {
-    /*
-    Update intra-morph references within a composite morph that has
-    been copied. For example, if a button refers to morph X in the
-    orginal composite then the copy of that button in the new composite
-    should refer to the copy of X in new composite, not the original X.
-    */
-    var property;
-    for (property in this) {
-        if (this[property] && this[property].isMorph && dict[property]) {
-            this[property] = dict[property];
-        }
-    }
-};
-
-// Morph dragging and dropping:
-
-Morph.prototype.rootForGrab = function () {
-    if (this instanceof ShadowMorph) {
-        return this.parent.rootForGrab();
-    }
-    if (this.parent instanceof ScrollFrameMorph) {
-        return this.parent;
-    }
-    if (this.parent === null ||
-            this.parent instanceof WorldMorph ||
-            this.parent instanceof FrameMorph ||
-            this.isDraggable === true) {
-        return this;
-    }
-    return this.parent.rootForGrab();
-};
-
-Morph.prototype.wantsDropOf = function (aMorph) {
-    // default is to answer the general flag - change for my heirs
-    if ((aMorph instanceof HandleMorph) ||
-            (aMorph instanceof MenuMorph) ||
-            (aMorph instanceof InspectorMorph)) {
-        return false;
-    }
-    return this.acceptsDrops;
-};
-
-Morph.prototype.pickUp = function (wrrld) {
-    var world = wrrld || this.world();
-    this.setPosition(
-        world.hand.position().subtract(
-            this.extent().floorDivideBy(2)
-        )
-    );
-    world.hand.grab(this);
-};
-
-Morph.prototype.isPickedUp = function () {
-    return this.parentThatIsA(HandMorph) !== null;
-};
-
-Morph.prototype.situation = function () {
-    // answer a dictionary specifying where I am right now, so
-    // I can slide back to it if I'm dropped somewhere else
-    if (this.parent) {
-        return {
-            origin: this.parent,
-            position: this.position().subtract(this.parent.position())
-        };
-    }
-    return null;
-};
-
-Morph.prototype.slideBackTo = function (situation, inSteps) {
-    var steps = inSteps || 5,
-        pos = situation.origin.position().add(situation.position),
-        xStep = -(this.left() - pos.x) / steps,
-        yStep = -(this.top() - pos.y) / steps,
-        stepCount = 0,
-        oldStep = this.step,
-        oldFps = this.fps,
-        myself = this;
-
-    this.fps = 0;
-    this.step = function () {
-        myself.fullChanged();
-        myself.silentMoveBy(new Point(xStep, yStep));
-        myself.fullChanged();
-        stepCount += 1;
-        if (stepCount === steps) {
-            situation.origin.add(myself);
-            if (situation.origin.reactToDropOf) {
-                situation.origin.reactToDropOf(myself);
-            }
-            myself.step = oldStep;
-            myself.fps = oldFps;
-        }
-    };
-};
-
-// Morph utilities:
-
-Morph.prototype.nop = function () {
-    nop();
-};
-
-Morph.prototype.resize = function () {
-    this.world().activeHandle = new HandleMorph(this);
-};
-
-Morph.prototype.move = function () {
-    this.world().activeHandle = new HandleMorph(
-        this,
-        null,
-        null,
-        null,
-        null,
-        'move'
-    );
-};
-
-Morph.prototype.hint = function (msg) {
-    var m, text;
-    text = msg;
-    if (msg) {
-        if (msg.toString) {
-            text = msg.toString();
-        }
-    } else {
-        text = 'NULL';
-    }
-    m = new MenuMorph(this, text);
-    m.isDraggable = true;
-    m.popUpCenteredAtHand(this.world());
-};
-
-Morph.prototype.inform = function (msg) {
-    var m, text;
-    text = msg;
-    if (msg) {
-        if (msg.toString) {
-            text = msg.toString();
-        }
-    } else {
-        text = 'NULL';
-    }
-    m = new MenuMorph(this, text);
-    m.addItem("Ok");
-    m.isDraggable = true;
-    m.popUpCenteredAtHand(this.world());
-};
-
-Morph.prototype.prompt = function (
-    msg,
-    callback,
-    environment,
-    defaultContents,
-    width,
-    floorNum,
-    ceilingNum,
-    isRounded
-) {
-    var menu, entryField, slider, isNumeric;
-    if (ceilingNum) {
-        isNumeric = true;
-    }
-    menu = new MenuMorph(
-        callback || null,
-        msg || '',
-        environment || null
-    );
-    entryField = new StringFieldMorph(
-        defaultContents || '',
-        width || 100,
-        MorphicPreferences.prompterFontSize,
-        MorphicPreferences.prompterFontName,
-        false,
-        false,
-        isNumeric
-    );
-    menu.items.push(entryField);
-    if (ceilingNum || MorphicPreferences.useSliderForInput) {
-        slider = new SliderMorph(
-            floorNum || 0,
-            ceilingNum,
-            parseFloat(defaultContents),
-            Math.floor((ceilingNum - floorNum) / 4),
-            'horizontal'
-        );
-        slider.alpha = 1;
-        slider.color = new Color(225, 225, 225);
-        slider.button.color = menu.borderColor;
-        slider.button.highlightColor = slider.button.color.copy();
-        slider.button.highlightColor.b += 100;
-        slider.button.pressColor = slider.button.color.copy();
-        slider.button.pressColor.b += 150;
-        slider.setHeight(MorphicPreferences.prompterSliderSize);
-        if (isRounded) {
-            slider.action = function (num) {
-                entryField.changed();
-                entryField.text.text = Math.round(num).toString();
-                entryField.text.drawNew();
-                entryField.text.changed();
-                entryField.text.edit();
-            };
-        } else {
-            slider.action = function (num) {
-                entryField.changed();
-                entryField.text.text = num.toString();
-                entryField.text.drawNew();
-                entryField.text.changed();
-            };
-        }
-        menu.items.push(slider);
-    }
-
-    menu.addLine(2);
-    menu.addItem('Ok', function () {
-        return entryField.string();
-    });
-    menu.addItem('Cancel', function () {
-        return null;
-    });
-    menu.isDraggable = true;
-    menu.popUpAtHand(this.world());
-    entryField.text.edit();
-};
-
-Morph.prototype.pickColor = function (
-    msg,
-    callback,
-    environment,
-    defaultContents
-) {
-    var menu, colorPicker;
-    menu = new MenuMorph(
-        callback || null,
-        msg || '',
-        environment || null
-    );
-    colorPicker = new ColorPickerMorph(defaultContents);
-    menu.items.push(colorPicker);
-    menu.addLine(2);
-    menu.addItem('Ok', function () {
-        return colorPicker.getChoice();
-    });
-    menu.addItem('Cancel', function () {
-        return null;
-    });
-    menu.isDraggable = true;
-    menu.popUpAtHand(this.world());
-};
-
-Morph.prototype.inspect = function (anotherObject) {
-    var world = this.world instanceof Function ?
-            this.world() : this.root() || this.world,
-        inspector,
-        inspectee = this;
-
-    if (anotherObject) {
-        inspectee = anotherObject;
-    }
-    inspector = new InspectorMorph(inspectee);
-    inspector.setPosition(world.hand.position());
-    inspector.keepWithin(world);
-    world.add(inspector);
-    inspector.changed();
-};
-
-// Morph menus:
-
-Morph.prototype.contextMenu = function () {
-    var world;
-
-    if (this.customContextMenu) {
-        return this.customContextMenu;
-    }
-    world = this.world instanceof Function ? this.world() : this.world;
-    if (world && world.isDevMode) {
-        if (this.parent === world) {
-            return this.developersMenu();
-        }
-        return this.hierarchyMenu();
-    }
-    return this.userMenu() ||
-        (this.parent && this.parent.userMenu());
-};
-
-Morph.prototype.hierarchyMenu = function () {
-    var parents = this.allParents(),
-        world = this.world instanceof Function ? this.world() : this.world,
-        menu = new MenuMorph(this, null);
-
-    parents.forEach(function (each) {
-        if (each.developersMenu && (each !== world)) {
-            menu.addItem(each.toString().slice(0, 50), function () {
-                each.developersMenu().popUpAtHand(world);
-            });
-        }
-    });
-    return menu;
-};
-
-Morph.prototype.developersMenu = function () {
-    // 'name' is not an official property of a function, hence:
-    var world = this.world instanceof Function ? this.world() : this.world,
-        userMenu = this.userMenu() ||
-            (this.parent && this.parent.userMenu()),
-        menu = new MenuMorph(this, this.constructor.name ||
-            this.constructor.toString().split(' ')[1].split('(')[0]);
-    if (userMenu) {
-        menu.addItem(
-            'user features...',
-            function () {
-                userMenu.popUpAtHand(world);
-            }
-        );
-        menu.addLine();
-    }
-    menu.addItem(
-        "color...",
-        function () {
-            this.pickColor(
-                menu.title + '\ncolor:',
-                this.setColor,
-                this,
-                this.color
-            );
-        },
-        'choose another color \nfor this morph'
-    );
-    menu.addItem(
-        "transparency...",
-        function () {
-            this.prompt(
-                menu.title + '\nalpha\nvalue:',
-                this.setAlphaScaled,
-                this,
-                (this.alpha * 100).toString(),
-                null,
-                1,
-                100,
-                true
-            );
-        },
-        'set this morph\'s\nalpha value'
-    );
-    menu.addItem(
-        "resize...",
-        'resize',
-        'show a handle\nwhich can be dragged\nto change this morph\'s' +
-            ' extent'
-    );
-    menu.addLine();
-    menu.addItem(
-        "duplicate",
-        function () {
-            this.fullCopy().pickUp(this.world());
-        },
-        'make a copy\nand pick it up'
-    );
-    menu.addItem(
-        "pick up",
-        'pickUp',
-        'disattach and put \ninto the hand'
-    );
-    menu.addItem(
-        "attach...",
-        'attach',
-        'stick this morph\nto another one'
-    );
-    menu.addItem(
-        "move...",
-        'move',
-        'show a handle\nwhich can be dragged\nto move this morph'
-    );
-    menu.addItem(
-        "inspect...",
-        'inspect',
-        'open a window\non all properties'
-    );
-    menu.addItem(
-        "pic...",
-        function () {
-            window.open(this.fullImageClassic().toDataURL());
-        },
-        'open a new window\nwith a picture of this morph'
-    );
-    menu.addLine();
-    if (this.isDraggable) {
-        menu.addItem(
-            "lock",
-            'toggleIsDraggable',
-            'make this morph\nunmovable'
-        );
-    } else {
-        menu.addItem(
-            "unlock",
-            'toggleIsDraggable',
-            'make this morph\nmovable'
-        );
-    }
-    menu.addItem("hide", 'hide');
-    menu.addItem("delete", 'destroy');
-    if (!(this instanceof WorldMorph)) {
-        menu.addLine();
-        menu.addItem(
-            "World...",
-            function () {
-                world.contextMenu().popUpAtHand(world);
-            },
-            'show the\nWorld\'s menu'
-        );
-    }
-    return menu;
-};
-
-Morph.prototype.userMenu = function () {
-    return null;
-};
-
-// Morph menu actions
-
-Morph.prototype.setAlphaScaled = function (alpha) {
-    // for context menu demo purposes
-    var newAlpha, unscaled;
-    if (typeof alpha === 'number') {
-        unscaled = alpha / 100;
-        this.alpha = Math.min(Math.max(unscaled, 0.1), 1);
-    } else {
-        newAlpha = parseFloat(alpha);
-        if (!isNaN(newAlpha)) {
-            unscaled = newAlpha / 100;
-            this.alpha = Math.min(Math.max(unscaled, 0.1), 1);
-        }
-    }
-    this.changed();
-};
-
-Morph.prototype.attach = function () {
-    var choices = this.overlappedMorphs(),
-        menu = new MenuMorph(this, 'choose new parent:'),
-        myself = this;
-
-    choices.forEach(function (each) {
-        menu.addItem(each.toString().slice(0, 50), function () {
-            each.add(myself);
-            myself.isDraggable = false;
-        });
-    });
-    if (choices.length > 0) {
-        menu.popUpAtHand(this.world());
-    }
-};
-
-Morph.prototype.toggleIsDraggable = function () {
-    // for context menu demo purposes
-    this.isDraggable = !this.isDraggable;
-};
-
-Morph.prototype.colorSetters = function () {
-    // for context menu demo purposes
-    return ['color'];
-};
-
-Morph.prototype.numericalSetters = function () {
-    // for context menu demo purposes
-    return [
-        'setLeft',
-        'setTop',
-        'setWidth',
-        'setHeight',
-        'setAlphaScaled'
-    ];
-};
-
-// Morph entry field tabbing:
-
-Morph.prototype.allEntryFields = function () {
-    return this.allChildren().filter(function (each) {
-        return each.isEditable &&
-            (each instanceof StringMorph ||
-                each instanceof TextMorph);
-    });
-};
-
-Morph.prototype.nextEntryField = function (current) {
-    var fields = this.allEntryFields(),
-        idx = fields.indexOf(current);
-    if (idx !== -1) {
-        if (fields.length > idx + 1) {
-            return fields[idx + 1];
-        }
-    }
-    return fields[0];
-};
-
-Morph.prototype.previousEntryField = function (current) {
-    var fields = this.allEntryFields(),
-        idx = fields.indexOf(current);
-    if (idx !== -1) {
-        if (idx > 0) {
-            return fields[idx - 1];
-        }
-        return fields[fields.length - 1];
-    }
-    return fields[0];
-};
-
-Morph.prototype.tab = function (editField) {
-/*
-    the <tab> key was pressed in one of my edit fields.
-    invoke my "nextTab()" function if it exists, else
-    propagate it up my owner chain.
-*/
-    if (this.nextTab) {
-        this.nextTab(editField);
-    } else if (this.parent) {
-        this.parent.tab(editField);
-    }
-};
-
-Morph.prototype.backTab = function (editField) {
-/*
-    the <back tab> key was pressed in one of my edit fields.
-    invoke my "previousTab()" function if it exists, else
-    propagate it up my owner chain.
-*/
-    if (this.previousTab) {
-        this.previousTab(editField);
-    } else if (this.parent) {
-        this.parent.backTab(editField);
-    }
-};
-
-/*
-    the following are examples of what the navigation methods should
-    look like. Insert these at the World level for fallback, and at lower
-    levels in the Morphic tree (e.g. dialog boxes) for a more fine-grained
-    control over the tabbing cycle.
-
-Morph.prototype.nextTab = function (editField) {
-    var next = this.nextEntryField(editField);
-    editField.clearSelection();
-    next.selectAll();
-    next.edit();
-};
-
-Morph.prototype.previousTab = function (editField) {
-    var prev = this.previousEntryField(editField);
-    editField.clearSelection();
-    prev.selectAll();
-    prev.edit();
-};
-
-*/
-
-// Morph events:
-
-Morph.prototype.escalateEvent = function (functionName, arg) {
-    var handler = this.parent;
-    while (!handler[functionName] && handler.parent !== null) {
-        handler = handler.parent;
-    }
-    if (handler[functionName]) {
-        handler[functionName](arg);
-    }
-};
-
-// Morph eval:
-
-Morph.prototype.evaluateString = function (code) {
-    var result;
-
-    try {
-        result = eval(code);
-        this.drawNew();
-        this.changed();
-    } catch (err) {
-        this.inform(err);
-    }
-    return result;
-};
-
-// Morph collision detection:
-
-Morph.prototype.isTouching = function (otherMorph) {
-    var oImg = this.overlappingImage(otherMorph),
-        data = oImg.getContext('2d')
-            .getImageData(1, 1, oImg.width, oImg.height)
-            .data;
-    return detect(
-        data,
-        function (each) {
-            return each !== 0;
-        }
-    ) !== null;
-};
-
-Morph.prototype.overlappingImage = function (otherMorph) {
-    var fb = this.fullBounds(),
-        otherFb = otherMorph.fullBounds(),
-        oRect = fb.intersect(otherFb),
-        oImg = newCanvas(oRect.extent()),
-        ctx = oImg.getContext('2d');
-    if (oRect.width() < 1 || oRect.height() < 1) {
-        return newCanvas(new Point(1, 1));
-    }
-    ctx.drawImage(
-        this.fullImage(),
-        oRect.origin.x - fb.origin.x,
-        oRect.origin.y - fb.origin.y
-    );
-    ctx.globalCompositeOperation = 'source-in';
-    ctx.drawImage(
-        otherMorph.fullImage(),
-        otherFb.origin.x - oRect.origin.x,
-        otherFb.origin.y - oRect.origin.y
-    );
-    return oImg;
-};
-
-// ShadowMorph /////////////////////////////////////////////////////////
-
-// ShadowMorph inherits from Morph:
-
-ShadowMorph.prototype = new Morph();
-ShadowMorph.prototype.constructor = ShadowMorph;
-ShadowMorph.uber = Morph.prototype;
-
-// ShadowMorph instance creation:
-
-function ShadowMorph() {
-    this.init();
-}
-
-// HandleMorph ////////////////////////////////////////////////////////
-
-// I am a resize / move handle that can be attached to any Morph
-
-// HandleMorph inherits from Morph:
-
-HandleMorph.prototype = new Morph();
-HandleMorph.prototype.constructor = HandleMorph;
-HandleMorph.uber = Morph.prototype;
-
-// HandleMorph instance creation:
-
-function HandleMorph(target, minX, minY, insetX, insetY, type) {
-    // if insetY is missing, it will be the same as insetX
-    this.init(target, minX, minY, insetX, insetY, type);
-}
-
-HandleMorph.prototype.init = function (
-    target,
-    minX,
-    minY,
-    insetX,
-    insetY,
-    type
-) {
-    var size = MorphicPreferences.handleSize;
-    this.target = target || null;
-    this.minExtent = new Point(minX || 0, minY || 0);
-    this.inset = new Point(insetX || 0, insetY || insetX || 0);
-    this.type =  type || 'resize'; // can also be 'move'
-    HandleMorph.uber.init.call(this);
-    this.color = new Color(255, 255, 255);
-    this.isDraggable = false;
-    this.noticesTransparentClick = true;
-    this.setExtent(new Point(size, size));
-};
-
-// HandleMorph drawing:
-
-HandleMorph.prototype.drawNew = function () {
-    this.normalImage = newCanvas(this.extent());
-    this.highlightImage = newCanvas(this.extent());
-    this.drawOnCanvas(
-        this.normalImage,
-        this.color,
-        new Color(100, 100, 100)
-    );
-    this.drawOnCanvas(
-        this.highlightImage,
-        new Color(100, 100, 255),
-        new Color(255, 255, 255)
-    );
-    this.image = this.normalImage;
-    if (this.target) {
-        this.setPosition(
-            this.target.bottomRight().subtract(
-                this.extent().add(this.inset)
-            )
-        );
-        this.target.add(this);
-        this.target.changed();
-    }
-};
-
-HandleMorph.prototype.drawOnCanvas = function (
-    aCanvas,
-    color,
-    shadowColor
-) {
-    var context = aCanvas.getContext('2d'),
-        p1,
-        p11,
-        p2,
-        p22,
-        i;
-
-    context.lineWidth = 1;
-    context.lineCap = 'round';
-
-    context.strokeStyle = color.toString();
-
-    if (this.type === 'move') {
-
-        p1 = this.bottomLeft().subtract(this.position());
-        p11 = p1.copy();
-        p2 = this.topRight().subtract(this.position());
-        p22 = p2.copy();
-
-        for (i = 0; i <= this.height(); i = i + 6) {
-            p11.y = p1.y - i;
-            p22.y = p2.y - i;
-
-            context.beginPath();
-            context.moveTo(p11.x, p11.y);
-            context.lineTo(p22.x, p22.y);
-            context.closePath();
-            context.stroke();
-        }
-    }
-
-    p1 = this.bottomLeft().subtract(this.position());
-    p11 = p1.copy();
-    p2 = this.topRight().subtract(this.position());
-    p22 = p2.copy();
-
-    for (i = 0; i <= this.width(); i = i + 6) {
-        p11.x = p1.x + i;
-        p22.x = p2.x + i;
-
-        context.beginPath();
-        context.moveTo(p11.x, p11.y);
-        context.lineTo(p22.x, p22.y);
-        context.closePath();
-        context.stroke();
-    }
-
-    context.strokeStyle = shadowColor.toString();
-
-    if (this.type === 'move') {
-
-        p1 = this.bottomLeft().subtract(this.position());
-        p11 = p1.copy();
-        p2 = this.topRight().subtract(this.position());
-        p22 = p2.copy();
-
-        for (i = -2; i <= this.height(); i = i + 6) {
-            p11.y = p1.y - i;
-            p22.y = p2.y - i;
-
-            context.beginPath();
-            context.moveTo(p11.x, p11.y);
-            context.lineTo(p22.x, p22.y);
-            context.closePath();
-            context.stroke();
-        }
-    }
-
-    p1 = this.bottomLeft().subtract(this.position());
-    p11 = p1.copy();
-    p2 = this.topRight().subtract(this.position());
-    p22 = p2.copy();
-
-    for (i = 2; i <= this.width(); i = i + 6) {
-        p11.x = p1.x + i;
-        p22.x = p2.x + i;
-
-        context.beginPath();
-        context.moveTo(p11.x, p11.y);
-        context.lineTo(p22.x, p22.y);
-        context.closePath();
-        context.stroke();
-    }
-};
-
-// HandleMorph stepping:
-
-HandleMorph.prototype.step = null;
-
-HandleMorph.prototype.mouseDownLeft = function (pos) {
-    var world = this.root(),
-        offset = pos.subtract(this.bounds.origin),
-        myself = this;
-
-    if (!this.target) {
-        return null;
-    }
-    this.step = function () {
-        var newPos, newExt;
-        if (world.hand.mouseButton) {
-            newPos = world.hand.bounds.origin.copy().subtract(offset);
-            if (this.type === 'resize') {
-                newExt = newPos.add(
-                    myself.extent().add(myself.inset)
-                ).subtract(myself.target.bounds.origin);
-                newExt = newExt.max(myself.minExtent);
-                myself.target.setExtent(newExt);
-
-                myself.setPosition(
-                    myself.target.bottomRight().subtract(
-                        myself.extent().add(myself.inset)
-                    )
-                );
-            } else { // type === 'move'
-                myself.target.setPosition(
-                    newPos.subtract(this.target.extent())
-                        .add(this.extent())
-                );
-            }
-        } else {
-            this.step = null;
-        }
-    };
-    if (!this.target.step) {
-        this.target.step = function () {
-            nop();
-        };
-    }
-};
-
-// HandleMorph dragging and dropping:
-
-HandleMorph.prototype.rootForGrab = function () {
-    return this;
-};
-
-// HandleMorph events:
-
-HandleMorph.prototype.mouseEnter = function () {
-    this.image = this.highlightImage;
-    this.changed();
-};
-
-HandleMorph.prototype.mouseLeave = function () {
-    this.image = this.normalImage;
-    this.changed();
-};
-
-// HandleMorph duplicating:
-
-HandleMorph.prototype.copyRecordingReferences = function (dict) {
-    // inherited, see comment in Morph
-    var c = HandleMorph.uber.copyRecordingReferences.call(
-        this,
-        dict
-    );
-    if (c.target && dict[this.target]) {
-        c.target = (dict[this.target]);
-    }
-    return c;
-};
-
-// HandleMorph menu:
-
-HandleMorph.prototype.attach = function () {
-    var choices = this.overlappedMorphs(),
-        menu = new MenuMorph(this, 'choose target:'),
-        myself = this;
-
-    choices.forEach(function (each) {
-        menu.addItem(each.toString().slice(0, 50), function () {
-            myself.isDraggable = false;
-            myself.target = each;
-            myself.drawNew();
-            myself.noticesTransparentClick = true;
-        });
-    });
-    if (choices.length > 0) {
-        menu.popUpAtHand(this.world());
-    }
-};
-
-// PenMorph ////////////////////////////////////////////////////////////
-
-// I am a simple LOGO-wise turtle.
-
-// PenMorph: referenced constructors
-
-var PenMorph;
-
-// PenMorph inherits from Morph:
-
-PenMorph.prototype = new Morph();
-PenMorph.prototype.constructor = PenMorph;
-PenMorph.uber = Morph.prototype;
-
-// PenMorph instance creation:
-
-function PenMorph() {
-    this.init();
-}
-
-PenMorph.prototype.init = function () {
-    var size = MorphicPreferences.handleSize * 4;
-
-    // additional properties:
-    this.isWarped = false; // internal optimization
-    this.heading = 0;
-    this.isDown = true;
-    this.size = 1;
-    this.wantsRedraw = false;
-    this.penPoint = 'tip'; // or 'center"
-
-    HandleMorph.uber.init.call(this);
-    this.setExtent(new Point(size, size));
-};
-
-// PenMorph updating - optimized for warping, i.e atomic recursion
-
-PenMorph.prototype.changed = function () {
-    if (this.isWarped === false) {
-        var w = this.root();
-        if (w instanceof WorldMorph) {
-            w.broken.push(this.visibleBounds().spread());
-        }
-        if (this.parent) {
-            this.parent.childChanged(this);
-        }
-    }
-};
-
-// PenMorph display:
-
-PenMorph.prototype.drawNew = function (facing) {
-/*
-    my orientation can be overridden with the "facing" parameter to
-    implement Scratch-style rotation styles
-
-*/
-    var context, start, dest, left, right, len,
-        direction = facing || this.heading;
-
-    if (this.isWarped) {
-        this.wantsRedraw = true;
-        return;
-    }
-    this.image = newCanvas(this.extent());
-    context = this.image.getContext('2d');
-    len = this.width() / 2;
-    start = this.center().subtract(this.bounds.origin);
-
-    if (this.penPoint === 'tip') {
-        dest = start.distanceAngle(len * 0.75, direction - 180);
-        left = start.distanceAngle(len, direction + 195);
-        right = start.distanceAngle(len, direction - 195);
-    } else { // 'middle'
-        dest = start.distanceAngle(len * 0.75, direction);
-        left = start.distanceAngle(len * 0.33, direction + 230);
-        right = start.distanceAngle(len * 0.33, direction - 230);
-    }
-
-    context.fillStyle = this.color.toString();
-    context.beginPath();
-
-    context.moveTo(start.x, start.y);
-    context.lineTo(left.x, left.y);
-    context.lineTo(dest.x, dest.y);
-    context.lineTo(right.x, right.y);
-
-    context.closePath();
-    context.strokeStyle = 'white';
-    context.lineWidth = 3;
-    context.stroke();
-    context.strokeStyle = 'black';
-    context.lineWidth = 1;
-    context.stroke();
-    context.fill();
-
-};
-
-// PenMorph access:
-
-PenMorph.prototype.setHeading = function (degrees) {
-    this.heading = parseFloat(degrees) % 360;
-    this.drawNew();
-    this.changed();
-};
-
-// PenMorph drawing:
-
-PenMorph.prototype.drawLine = function (start, dest) {
-    var context = this.parent.penTrails().getContext('2d'),
-        from = start.subtract(this.parent.bounds.origin),
-        to = dest.subtract(this.parent.bounds.origin);
-    if (this.isDown) {
-        context.lineWidth = this.size;
-        context.strokeStyle = this.color.toString();
-        context.lineCap = 'round';
-        context.lineJoin = 'round';
-        context.beginPath();
-        context.moveTo(from.x, from.y);
-        context.lineTo(to.x, to.y);
-        context.stroke();
-        if (this.isWarped === false) {
-            this.world().broken.push(
-                start.rectangle(dest).expandBy(
-                    Math.max(this.size / 2, 1)
-                ).intersect(this.parent.visibleBounds()).spread()
-            );
-        }
-    }
-};
-
-// PenMorph turtle ops:
-
-PenMorph.prototype.turn = function (degrees) {
-    this.setHeading(this.heading + parseFloat(degrees));
-};
-
-PenMorph.prototype.forward = function (steps) {
-    var start = this.center(),
-        dest,
-        dist = parseFloat(steps);
-    if (dist >= 0) {
-        dest = this.position().distanceAngle(dist, this.heading);
-    } else {
-        dest = this.position().distanceAngle(
-            Math.abs(dist),
-            (this.heading - 180)
-        );
-    }
-    this.setPosition(dest);
-    this.drawLine(start, this.center());
-};
-
-PenMorph.prototype.down = function () {
-    this.isDown = true;
-};
-
-PenMorph.prototype.up = function () {
-    this.isDown = false;
-};
-
-PenMorph.prototype.clear = function () {
-    this.parent.drawNew();
-    this.parent.changed();
-};
-
-// PenMorph optimization for atomic recursion:
-
-PenMorph.prototype.startWarp = function () {
-    this.wantsRedraw = false;
-    this.isWarped = true;
-};
-
-PenMorph.prototype.endWarp = function () {
-    this.isWarped = false;
-    if (this.wantsRedraw) {
-        this.drawNew();
-        this.wantsRedraw = false;
-    }
-    this.parent.changed();
-};
-
-PenMorph.prototype.warp = function (fun) {
-    this.startWarp();
-    fun.call(this);
-    this.endWarp();
-};
-
-PenMorph.prototype.warpOp = function (selector, argsArray) {
-    this.startWarp();
-    this[selector].apply(this, argsArray);
-    this.endWarp();
-};
-
-// PenMorph demo ops:
-// try these with WARP eg.: this.warp(function () {tree(12, 120, 20)})
-
-PenMorph.prototype.warpSierpinski = function (length, min) {
-    this.warpOp('sierpinski', [length, min]);
-};
-
-PenMorph.prototype.sierpinski = function (length, min) {
-    var i;
-    if (length > min) {
-        for (i = 0; i < 3; i += 1) {
-            this.sierpinski(length * 0.5, min);
-            this.turn(120);
-            this.forward(length);
-        }
-    }
-};
-
-PenMorph.prototype.warpTree = function (level, length, angle) {
-    this.warpOp('tree', [level, length, angle]);
-};
-
-PenMorph.prototype.tree = function (level, length, angle) {
-    if (level > 0) {
-        this.size = level;
-        this.forward(length);
-        this.turn(angle);
-        this.tree(level - 1, length * 0.75, angle);
-        this.turn(angle * -2);
-        this.tree(level - 1, length * 0.75, angle);
-        this.turn(angle);
-        this.forward(-length);
-    }
-};
-
-// ColorPaletteMorph ///////////////////////////////////////////////////
-
-var ColorPaletteMorph;
-
-// ColorPaletteMorph inherits from Morph:
-
-ColorPaletteMorph.prototype = new Morph();
-ColorPaletteMorph.prototype.constructor = ColorPaletteMorph;
-ColorPaletteMorph.uber = Morph.prototype;
-
-// ColorPaletteMorph instance creation:
-
-function ColorPaletteMorph(target, sizePoint) {
-    this.init(
-        target || null,
-        sizePoint || new Point(80, 50)
-    );
-}
-
-ColorPaletteMorph.prototype.init = function (target, size) {
-    ColorPaletteMorph.uber.init.call(this);
-    this.target = target;
-    this.targetSetter = 'color';
-    this.silentSetExtent(size);
-    this.choice = null;
-    this.drawNew();
-};
-
-ColorPaletteMorph.prototype.drawNew = function () {
-    var context, ext, x, y, h, l;
-
-    ext = this.extent();
-    this.image = newCanvas(this.extent());
-    context = this.image.getContext('2d');
-    this.choice = new Color();
-    for (x = 0; x <= ext.x; x += 1) {
-        h = 360 * x / ext.x;
-        for (y = 0; y <= ext.y; y += 1) {
-            l = 100 - (y / ext.y * 100);
-            context.fillStyle = 'hsl(' + h + ',100%,' + l + '%)';
-            context.fillRect(x, y, 1, 1);
-        }
-    }
-};
-
-ColorPaletteMorph.prototype.mouseMove = function (pos) {
-    this.choice = this.getPixelColor(pos);
-    this.updateTarget();
-};
-
-ColorPaletteMorph.prototype.mouseDownLeft = function (pos) {
-    this.choice = this.getPixelColor(pos);
-    this.updateTarget();
-};
-
-ColorPaletteMorph.prototype.updateTarget = function () {
-    if (this.target instanceof Morph && this.choice !== null) {
-        if (this.target[this.targetSetter] instanceof Function) {
-            this.target[this.targetSetter](this.choice);
-        } else {
-            this.target[this.targetSetter] = this.choice;
-            this.target.drawNew();
-            this.target.changed();
-        }
-    }
-};
-
-// ColorPaletteMorph duplicating:
-
-ColorPaletteMorph.prototype.copyRecordingReferences = function (dict) {
-    // inherited, see comment in Morph
-    var c = ColorPaletteMorph.uber.copyRecordingReferences.call(
-        this,
-        dict
-    );
-    if (c.target && dict[this.target]) {
-        c.target = (dict[this.target]);
-    }
-    return c;
-};
-
-// ColorPaletteMorph menu:
-
-ColorPaletteMorph.prototype.developersMenu = function () {
-    var menu = ColorPaletteMorph.uber.developersMenu.call(this);
-    menu.addLine();
-    menu.addItem(
-        'set target',
-        "setTarget",
-        'choose another morph\nwhose color property\n will be' +
-            ' controlled by this one'
-    );
-    return menu;
-};
-
-ColorPaletteMorph.prototype.setTarget = function () {
-    var choices = this.overlappedMorphs(),
-        menu = new MenuMorph(this, 'choose target:'),
-        myself = this;
-
-    choices.push(this.world());
-    choices.forEach(function (each) {
-        menu.addItem(each.toString().slice(0, 50), function () {
-            myself.target = each;
-            myself.setTargetSetter();
-        });
-    });
-    if (choices.length === 1) {
-        this.target = choices[0];
-        this.setTargetSetter();
-    } else if (choices.length > 0) {
-        menu.popUpAtHand(this.world());
-    }
-};
-
-ColorPaletteMorph.prototype.setTargetSetter = function () {
-    var choices = this.target.colorSetters(),
-        menu = new MenuMorph(this, 'choose target property:'),
-        myself = this;
-
-    choices.forEach(function (each) {
-        menu.addItem(each, function () {
-            myself.targetSetter = each;
-        });
-    });
-    if (choices.length === 1) {
-        this.targetSetter = choices[0];
-    } else if (choices.length > 0) {
-        menu.popUpAtHand(this.world());
-    }
-};
-
-// GrayPaletteMorph ///////////////////////////////////////////////////
-
-var GrayPaletteMorph;
-
-// GrayPaletteMorph inherits from ColorPaletteMorph:
-
-GrayPaletteMorph.prototype = new ColorPaletteMorph();
-GrayPaletteMorph.prototype.constructor = GrayPaletteMorph;
-GrayPaletteMorph.uber = ColorPaletteMorph.prototype;
-
-// GrayPaletteMorph instance creation:
-
-function GrayPaletteMorph(target, sizePoint) {
-    this.init(
-        target || null,
-        sizePoint || new Point(80, 10)
-    );
-}
-
-GrayPaletteMorph.prototype.drawNew = function () {
-    var context, ext, gradient;
-
-    ext = this.extent();
-    this.image = newCanvas(this.extent());
-    context = this.image.getContext('2d');
-    this.choice = new Color();
-    gradient = context.createLinearGradient(0, 0, ext.x, ext.y);
-    gradient.addColorStop(0, 'black');
-    gradient.addColorStop(1, 'white');
-    context.fillStyle = gradient;
-    context.fillRect(0, 0, ext.x, ext.y);
-};
-
-// ColorPickerMorph ///////////////////////////////////////////////////
-
-// ColorPickerMorph inherits from Morph:
-
-ColorPickerMorph.prototype = new Morph();
-ColorPickerMorph.prototype.constructor = ColorPickerMorph;
-ColorPickerMorph.uber = Morph.prototype;
-
-// ColorPickerMorph instance creation:
-
-function ColorPickerMorph(defaultColor) {
-    this.init(defaultColor || new Color(255, 255, 255));
-}
-
-ColorPickerMorph.prototype.init = function (defaultColor) {
-    this.choice = defaultColor;
-    ColorPickerMorph.uber.init.call(this);
-    this.color = new Color(255, 255, 255);
-    this.silentSetExtent(new Point(80, 80));
-    this.drawNew();
-};
-
-ColorPickerMorph.prototype.drawNew = function () {
-    ColorPickerMorph.uber.drawNew.call(this);
-    this.buildSubmorphs();
-};
-
-ColorPickerMorph.prototype.buildSubmorphs = function () {
-    var cpal, gpal, x, y;
-
-    this.children.forEach(function (child) {
-        child.destroy();
-    });
-    this.children = [];
-    this.feedback = new Morph();
-    this.feedback.color = this.choice;
-    this.feedback.setExtent(new Point(20, 20));
-    cpal = new ColorPaletteMorph(
-        this.feedback,
-        new Point(this.width(), 50)
-    );
-    gpal = new GrayPaletteMorph(
-        this.feedback,
-        new Point(this.width(), 5)
-    );
-    cpal.setPosition(this.bounds.origin);
-    this.add(cpal);
-    gpal.setPosition(cpal.bottomLeft());
-    this.add(gpal);
-    x = (gpal.left() +
-        Math.floor((gpal.width() - this.feedback.width()) / 2));
-    y = gpal.bottom() + Math.floor((this.bottom() -
-        gpal.bottom() - this.feedback.height()) / 2);
-    this.feedback.setPosition(new Point(x, y));
-    this.add(this.feedback);
-};
-
-ColorPickerMorph.prototype.getChoice = function () {
-    return this.feedback.color;
-};
-
-ColorPickerMorph.prototype.rootForGrab = function () {
-    return this;
-};
-
-// BlinkerMorph ////////////////////////////////////////////////////////
-
-// can be used for text cursors
-
-var BlinkerMorph;
-
-// BlinkerMorph inherits from Morph:
-
-BlinkerMorph.prototype = new Morph();
-BlinkerMorph.prototype.constructor = BlinkerMorph;
-BlinkerMorph.uber = Morph.prototype;
-
-// BlinkerMorph instance creation:
-
-function BlinkerMorph(rate) {
-    this.init(rate);
-}
-
-BlinkerMorph.prototype.init = function (rate) {
-    BlinkerMorph.uber.init.call(this);
-    this.color = new Color(0, 0, 0);
-    this.fps = rate || 2;
-    this.drawNew();
-};
-
-// BlinkerMorph stepping:
-
-BlinkerMorph.prototype.step = function () {
-    this.toggleVisibility();
-};
-
-// CursorMorph /////////////////////////////////////////////////////////
-
-// I am a String/Text editing widget
-
-// CursorMorph: referenced constructors
-
-var CursorMorph;
-
-// CursorMorph inherits from BlinkerMorph:
-
-CursorMorph.prototype = new BlinkerMorph();
-CursorMorph.prototype.constructor = CursorMorph;
-CursorMorph.uber = BlinkerMorph.prototype;
-
-// CursorMorph preferences settings:
-
-CursorMorph.prototype.viewPadding = 1;
-
-// CursorMorph instance creation:
-
-function CursorMorph(aStringOrTextMorph) {
-    this.init(aStringOrTextMorph);
-}
-
-CursorMorph.prototype.init = function (aStringOrTextMorph) {
-    var ls;
-
-    // additional properties:
-    this.keyDownEventUsed = false;
-    this.target = aStringOrTextMorph;
-    this.originalContents = this.target.text;
-    this.originalAlignment = this.target.alignment;
-    this.slot = this.target.text.length;
-    CursorMorph.uber.init.call(this);
-    ls = fontHeight(this.target.fontSize);
-    this.setExtent(new Point(Math.max(Math.floor(ls / 20), 1), ls));
-    this.drawNew();
-    this.image.getContext('2d').font = this.target.font();
-    if (this.target instanceof TextMorph &&
-            (this.target.alignment !== 'left')) {
-        this.target.setAlignmentToLeft();
-    }
-    this.gotoSlot(this.slot);
-};
-
-// CursorMorph event processing:
-
-CursorMorph.prototype.processKeyPress = function (event) {
-    // this.inspectKeyEvent(event);
-    if (this.keyDownEventUsed) {
-        this.keyDownEventUsed = false;
-        return null;
-    }
-    if ((event.keyCode === 40) || event.charCode === 40) {
-        this.insert('(');
-        return null;
-    }
-    if ((event.keyCode === 37) || event.charCode === 37) {
-        this.insert('%');
-        return null;
-    }
-    if (event.keyCode) { // Opera doesn't support charCode
-        if (event.ctrlKey) {
-            this.ctrl(event.keyCode, event.shiftKey);
-        } else if (event.metaKey) {
-            this.cmd(event.keyCode, event.shiftKey);
-        } else {
-            this.insert(
-                String.fromCharCode(event.keyCode),
-                event.shiftKey
-            );
-        }
-    } else if (event.charCode) { // all other browsers
-        if (event.ctrlKey) {
-            this.ctrl(event.charCode, event.shiftKey);
-        } else if (event.metaKey) {
-            this.cmd(event.charCode, event.shiftKey);
-        } else {
-            this.insert(
-                String.fromCharCode(event.charCode),
-                event.shiftKey
-            );
-        }
-    }
-    // notify target's parent of key event
-    this.target.escalateEvent('reactToKeystroke', event);
-};
-
-CursorMorph.prototype.processKeyDown = function (event) {
-    // this.inspectKeyEvent(event);
-    var shift = event.shiftKey;
-    this.keyDownEventUsed = false;
-    if (event.ctrlKey) {
-        this.ctrl(event.keyCode, event.shiftKey);
-        // notify target's parent of key event
-        this.target.escalateEvent('reactToKeystroke', event);
-        return;
-    }
-    if (event.metaKey) {
-        this.cmd(event.keyCode, event.shiftKey);
-        // notify target's parent of key event
-        this.target.escalateEvent('reactToKeystroke', event);
-        return;
-    }
-
-    switch (event.keyCode) {
-    case 37:
-        this.goLeft(shift);
-        this.keyDownEventUsed = true;
-        break;
-    case 39:
-        this.goRight(shift);
-        this.keyDownEventUsed = true;
-        break;
-    case 38:
-        this.goUp(shift);
-        this.keyDownEventUsed = true;
-        break;
-    case 40:
-        this.goDown(shift);
-        this.keyDownEventUsed = true;
-        break;
-    case 36:
-        this.goHome(shift);
-        this.keyDownEventUsed = true;
-        break;
-    case 35:
-        this.goEnd(shift);
-        this.keyDownEventUsed = true;
-        break;
-    case 46:
-        this.deleteRight();
-        this.keyDownEventUsed = true;
-        break;
-    case 8:
-        this.deleteLeft();
-        this.keyDownEventUsed = true;
-        break;
-    case 13:
-        if (this.target instanceof StringMorph) {
-            this.accept();
-        } else {
-            this.insert('\n');
-        }
-        this.keyDownEventUsed = true;
-        break;
-    case 27:
-        this.cancel();
-        this.keyDownEventUsed = true;
-        break;
-    default:
-        nop();
-        // this.inspectKeyEvent(event);
-    }
-    // notify target's parent of key event
-    this.target.escalateEvent('reactToKeystroke', event);
-};
-
-// CursorMorph navigation:
-
-/*
-// original non-scrolling code, commented out in case we need to fall back:
-
-CursorMorph.prototype.gotoSlot = function (newSlot) {
-    this.setPosition(this.target.slotPosition(newSlot));
-    this.slot = Math.max(newSlot, 0);
-};
-*/
-
-CursorMorph.prototype.gotoSlot = function (slot) {
-    var length = this.target.text.length,
-        pos = this.target.slotPosition(slot),
-        right,
-        left;
-    this.slot = slot < 0 ? 0 : slot > length ? length : slot;
-    if (this.parent && this.target.isScrollable) {
-        right = this.parent.right() - this.viewPadding;
-        left = this.parent.left() + this.viewPadding;
-        if (pos.x > right) {
-            this.target.setLeft(this.target.left() + right - pos.x);
-            pos.x = right;
-        }
-        if (pos.x < left) {
-            left = Math.min(this.parent.left(), left);
-            this.target.setLeft(this.target.left() + left - pos.x);
-            pos.x = left;
-        }
-        if (this.target.right() < right &&
-                right - this.target.width() < left) {
-            pos.x += right - this.target.right();
-            this.target.setRight(right);
-        }
-    }
-    this.show();
-    this.setPosition(pos);
-    if (this.parent
-            && this.parent.parent instanceof ScrollFrameMorph
-            && this.target.isScrollable) {
-        this.parent.parent.scrollCursorIntoView(this);
-    }
-};
-
-CursorMorph.prototype.goLeft = function (shift) {
-    this.updateSelection(shift);
-    this.gotoSlot(this.slot - 1);
-    this.updateSelection(shift);
-};
-
-CursorMorph.prototype.goRight = function (shift, howMany) {
-    this.updateSelection(shift);
-    this.gotoSlot(this.slot + (howMany || 1));
-    this.updateSelection(shift);
-};
-
-CursorMorph.prototype.goUp = function (shift) {
-    this.updateSelection(shift);
-    this.gotoSlot(this.target.upFrom(this.slot));
-    this.updateSelection(shift);
-};
-
-CursorMorph.prototype.goDown = function (shift) {
-    this.updateSelection(shift);
-    this.gotoSlot(this.target.downFrom(this.slot));
-    this.updateSelection(shift);
-};
-
-CursorMorph.prototype.goHome = function (shift) {
-    this.updateSelection(shift);
-    this.gotoSlot(this.target.startOfLine(this.slot));
-    this.updateSelection(shift);
-};
-
-CursorMorph.prototype.goEnd = function (shift) {
-    this.updateSelection(shift);
-    this.gotoSlot(this.target.endOfLine(this.slot));
-    this.updateSelection(shift);
-};
-
-CursorMorph.prototype.gotoPos = function (aPoint) {
-    this.gotoSlot(this.target.slotAt(aPoint));
-    this.show();
-};
-
-// CursorMorph selecting:
-
-CursorMorph.prototype.updateSelection = function (shift) {
-    if (shift) {
-        if (!this.target.endMark && !this.target.startMark) {
-            this.target.startMark = this.slot;
-            this.target.endMark = this.slot;
-        } else if (this.target.endMark !== this.slot) {
-            this.target.endMark = this.slot;
-            this.target.drawNew();
-            this.target.changed();
-        }
-    } else {
-        this.target.clearSelection();
-    }
-};
-
-// CursorMorph editing:
-
-CursorMorph.prototype.accept = function () {
-    var world = this.root();
-    if (world) {
-        world.stopEditing();
-    }
-    this.escalateEvent('accept', null);
-};
-
-CursorMorph.prototype.cancel = function () {
-    var world = this.root();
-    this.undo();
-    if (world) {
-        world.stopEditing();
-    }
-    this.escalateEvent('cancel', null);
-};
-
-CursorMorph.prototype.undo = function () {
-    this.target.text = this.originalContents;
-    this.target.changed();
-    this.target.drawNew();
-    this.target.changed();
-    this.gotoSlot(0);
-};
-
-CursorMorph.prototype.insert = function (aChar, shiftKey) {
-    var text;
-    if (aChar === '\u0009') {
-        this.target.escalateEvent('reactToEdit', this.target);
-        if (shiftKey) {
-            return this.target.backTab(this.target);
-        }
-        return this.target.tab(this.target);
-    }
-    if (!this.target.isNumeric ||
-            !isNaN(parseFloat(aChar)) ||
-            contains(['-', '.'], aChar)) {
-        if (this.target.selection() !== '') {
-            this.gotoSlot(this.target.selectionStartSlot());
-            this.target.deleteSelection();
-        }
-        text = this.target.text;
-        text = text.slice(0, this.slot) +
-            aChar +
-            text.slice(this.slot);
-        this.target.text = text;
-        this.target.drawNew();
-        this.target.changed();
-        this.goRight(false, aChar.length);
-    }
-};
-
-CursorMorph.prototype.ctrl = function (aChar, shiftKey) {
-    if (aChar === 64 || (aChar === 65 && shiftKey)) {
-        this.insert('@');
-    } else if (aChar === 65) {
-        this.target.selectAll();
-    } else if (aChar === 90) {
-        this.undo();
-    } else if (aChar === 123) {
-        this.insert('{');
-    } else if (aChar === 125) {
-        this.insert('}');
-    } else if (aChar === 91) {
-        this.insert('[');
-    } else if (aChar === 93) {
-        this.insert(']');
-    } else if (!isNil(this.target.receiver)) {
-        if (aChar === 68) {
-            this.target.doIt();
-        } else if (aChar === 73) {
-            this.target.inspectIt();
-        } else if (aChar === 80) {
-            this.target.showIt();
-        }
-    }
-
-
-};
-
-CursorMorph.prototype.cmd = function (aChar, shiftKey) {
-    if (aChar === 64 || (aChar === 65 && shiftKey)) {
-        this.insert('@');
-    } else if (aChar === 65) {
-        this.target.selectAll();
-    } else if (aChar === 90) {
-        this.undo();
-    } else if (!isNil(this.target.receiver)) {
-        if (aChar === 68) {
-            this.target.doIt();
-        } else if (aChar === 73) {
-            this.target.inspectIt();
-        } else if (aChar === 80) {
-            this.target.showIt();
-        }
-    }
-};
-
-CursorMorph.prototype.deleteRight = function () {
-    var text;
-    if (this.target.selection() !== '') {
-        this.gotoSlot(this.target.selectionStartSlot());
-        this.target.deleteSelection();
-    } else {
-        text = this.target.text;
-        this.target.changed();
-        text = text.slice(0, this.slot) + text.slice(this.slot + 1);
-        this.target.text = text;
-        this.target.drawNew();
-    }
-};
-
-CursorMorph.prototype.deleteLeft = function () {
-    var text;
-    if (this.target.selection()) {
-        this.gotoSlot(this.target.selectionStartSlot());
-        return this.target.deleteSelection();
-    }
-    text = this.target.text;
-    this.target.changed();
-    this.target.text = text.substring(0, this.slot - 1) +
-        text.substr(this.slot);
-    this.target.drawNew();
-    this.goLeft();
-};
-
-// CursorMorph destroying:
-
-CursorMorph.prototype.destroy = function () {
-    if (this.target.alignment !== this.originalAlignment) {
-        this.target.alignment = this.originalAlignment;
-        this.target.drawNew();
-        this.target.changed();
-    }
-    CursorMorph.uber.destroy.call(this);
-};
-
-// CursorMorph utilities:
-
-CursorMorph.prototype.inspectKeyEvent = function (event) {
-    // private
-    this.inform(
-        'Key pressed: ' +
-            String.fromCharCode(event.charCode) +
-            '\n------------------------' +
-            '\ncharCode: ' +
-            event.charCode.toString() +
-            '\nkeyCode: ' +
-            event.keyCode.toString() +
-            '\nshiftKey: ' +
-            event.shiftKey.toString() +
-            '\naltKey: ' +
-            event.altKey.toString() +
-            '\nctrlKey: ' +
-            event.ctrlKey.toString() +
-            '\ncmdKey: ' +
-            event.metaKey.toString()
-    );
-};
-
-// BoxMorph ////////////////////////////////////////////////////////////
-
-// I can have an optionally rounded border
-
-var BoxMorph;
-
-// BoxMorph inherits from Morph:
-
-BoxMorph.prototype = new Morph();
-BoxMorph.prototype.constructor = BoxMorph;
-BoxMorph.uber = Morph.prototype;
-
-// BoxMorph instance creation:
-
-function BoxMorph(edge, border, borderColor) {
-    this.init(edge, border, borderColor);
-}
-
-BoxMorph.prototype.init = function (edge, border, borderColor) {
-    this.edge = edge || 4;
-    this.border = border || ((border === 0) ? 0 : 2);
-    this.borderColor = borderColor || new Color();
-    BoxMorph.uber.init.call(this);
-};
-
-// BoxMorph drawing:
-
-BoxMorph.prototype.drawNew = function () {
-    var context;
-
-    this.image = newCanvas(this.extent());
-    context = this.image.getContext('2d');
-    if ((this.edge === 0) && (this.border === 0)) {
-        BoxMorph.uber.drawNew.call(this);
-        return null;
-    }
-    context.fillStyle = this.color.toString();
-    context.beginPath();
-    this.outlinePath(
-        context,
-        Math.max(this.edge - this.border, 0),
-        this.border
-    );
-    context.closePath();
-    context.fill();
-    if (this.border > 0) {
-        context.lineWidth = this.border;
-        context.strokeStyle = this.borderColor.toString();
-        context.beginPath();
-        this.outlinePath(context, this.edge, this.border / 2);
-        context.closePath();
-        context.stroke();
-    }
-};
-
-BoxMorph.prototype.outlinePath = function (context, radius, inset) {
-    var offset = radius + inset,
-        w = this.width(),
-        h = this.height();
-
-    // top left:
-    context.arc(
-        offset,
-        offset,
-        radius,
-        radians(-180),
-        radians(-90),
-        false
-    );
-    // top right:
-    context.arc(
-        w - offset,
-        offset,
-        radius,
-        radians(-90),
-        radians(-0),
-        false
-    );
-    // bottom right:
-    context.arc(
-        w - offset,
-        h - offset,
-        radius,
-        radians(0),
-        radians(90),
-        false
-    );
-    // bottom left:
-    context.arc(
-        offset,
-        h - offset,
-        radius,
-        radians(90),
-        radians(180),
-        false
-    );
-};
-
-
-// BoxMorph menus:
-
-BoxMorph.prototype.developersMenu = function () {
-    var menu = BoxMorph.uber.developersMenu.call(this);
-    menu.addLine();
-    menu.addItem(
-        "border width...",
-        function () {
-            this.prompt(
-                menu.title + '\nborder\nwidth:',
-                this.setBorderWidth,
-                this,
-                this.border.toString(),
-                null,
-                0,
-                100,
-                true
-            );
-        },
-        'set the border\'s\nline size'
-    );
-    menu.addItem(
-        "border color...",
-        function () {
-            this.pickColor(
-                menu.title + '\nborder color:',
-                this.setBorderColor,
-                this,
-                this.borderColor
-            );
-        },
-        'set the border\'s\nline color'
-    );
-    menu.addItem(
-        "corner size...",
-        function () {
-            this.prompt(
-                menu.title + '\ncorner\nsize:',
-                this.setCornerSize,
-                this,
-                this.edge.toString(),
-                null,
-                0,
-                100,
-                true
-            );
-        },
-        'set the corner\'s\nradius'
-    );
-    return menu;
-};
-
-BoxMorph.prototype.setBorderWidth = function (size) {
-    // for context menu demo purposes
-    var newSize;
-    if (typeof size === 'number') {
-        this.border = Math.max(size, 0);
-    } else {
-        newSize = parseFloat(size);
-        if (!isNaN(newSize)) {
-            this.border = Math.max(newSize, 0);
-        }
-    }
-    this.drawNew();
-    this.changed();
-};
-
-BoxMorph.prototype.setBorderColor = function (color) {
-    // for context menu demo purposes
-    if (color) {
-        this.borderColor = color;
-        this.drawNew();
-        this.changed();
-    }
-};
-
-BoxMorph.prototype.setCornerSize = function (size) {
-    // for context menu demo purposes
-    var newSize;
-    if (typeof size === 'number') {
-        this.edge = Math.max(size, 0);
-    } else {
-        newSize = parseFloat(size);
-        if (!isNaN(newSize)) {
-            this.edge = Math.max(newSize, 0);
-        }
-    }
-    this.drawNew();
-    this.changed();
-};
-
-BoxMorph.prototype.colorSetters = function () {
-    // for context menu demo purposes
-    return ['color', 'borderColor'];
-};
-
-BoxMorph.prototype.numericalSetters = function () {
-    // for context menu demo purposes
-    var list = BoxMorph.uber.numericalSetters.call(this);
-    list.push('setBorderWidth', 'setCornerSize');
-    return list;
-};
-
-// SpeechBubbleMorph ///////////////////////////////////////////////////
-
-/*
-    I am a comic-style speech bubble that can display either a string,
-    a Morph, a Canvas or a toString() representation of anything else.
-    If I am invoked using popUp() I behave like a tool tip.
-*/
-
-// SpeechBubbleMorph: referenced constructors
-
-var SpeechBubbleMorph;
-
-// SpeechBubbleMorph inherits from BoxMorph:
-
-SpeechBubbleMorph.prototype = new BoxMorph();
-SpeechBubbleMorph.prototype.constructor = SpeechBubbleMorph;
-SpeechBubbleMorph.uber = BoxMorph.prototype;
-
-// SpeechBubbleMorph instance creation:
-
-function SpeechBubbleMorph(
-    contents,
-    color,
-    edge,
-    border,
-    borderColor,
-    padding,
-    isThought
-) {
-    this.init(contents, color, edge, border, borderColor, padding, isThought);
-}
-
-SpeechBubbleMorph.prototype.init = function (
-    contents,
-    color,
-    edge,
-    border,
-    borderColor,
-    padding,
-    isThought
-) {
-    this.isPointingRight = true; // orientation of text
-    this.contents = contents || '';
-    this.padding = padding || 0; // additional vertical pixels
-    this.isThought = isThought || false; // draw "think" bubble
-    this.isClickable = false;
-    SpeechBubbleMorph.uber.init.call(
-        this,
-        edge || 6,
-        border || ((border === 0) ? 0 : 1),
-        borderColor || new Color(140, 140, 140)
-    );
-    this.color = color || new Color(230, 230, 230);
-    this.drawNew();
-};
-
-// SpeechBubbleMorph invoking:
-
-SpeechBubbleMorph.prototype.popUp = function (world, pos, isClickable) {
-    this.drawNew();
-    this.setPosition(pos.subtract(new Point(0, this.height())));
-    this.addShadow(new Point(2, 2), 80);
-    this.keepWithin(world);
-    world.add(this);
-    this.fullChanged();
-    world.hand.destroyTemporaries();
-    world.hand.temporaries.push(this);
-
-    if (!isClickable) {
-        this.mouseEnter = function () {
-            this.destroy();
-        };
-    } else {
-        this.isClickable = true;
-    }
-};
-
-// SpeechBubbleMorph drawing:
-
-SpeechBubbleMorph.prototype.drawNew = function () {
-    // re-build my contents
-    if (this.contentsMorph) {
-        this.contentsMorph.destroy();
-    }
-    if (this.contents instanceof Morph) {
-        this.contentsMorph = this.contents;
-    } else if (isString(this.contents)) {
-        this.contentsMorph = new TextMorph(
-            this.contents,
-            MorphicPreferences.bubbleHelpFontSize,
-            null,
-            false,
-            true,
-            'center'
-        );
-    } else if (this.contents instanceof HTMLCanvasElement) {
-        this.contentsMorph = new Morph();
-        this.contentsMorph.silentSetWidth(this.contents.width);
-        this.contentsMorph.silentSetHeight(this.contents.height);
-        this.contentsMorph.image = this.contents;
-    } else {
-        this.contentsMorph = new TextMorph(
-            this.contents.toString(),
-            MorphicPreferences.bubbleHelpFontSize,
-            null,
-            false,
-            true,
-            'center'
-        );
-    }
-    this.add(this.contentsMorph);
-
-    // adjust my layout
-    this.silentSetWidth(this.contentsMorph.width() +
-        (this.padding ? this.padding * 2 : this.edge * 2));
-    this.silentSetHeight(this.contentsMorph.height() +
-        this.edge +
-        this.border * 2 +
-        this.padding * 2 +
-        2);
-
-    // draw my outline
-    SpeechBubbleMorph.uber.drawNew.call(this);
-
-    // position my contents
-    this.contentsMorph.setPosition(this.position().add(
-        new Point(
-            this.padding || this.edge,
-            this.border + this.padding + 1
-        )
-    ));
-};
-
-SpeechBubbleMorph.prototype.outlinePath = function (
-    context,
-    radius,
-    inset
-) {
-    var offset = radius + inset,
-        w = this.width(),
-        h = this.height(),
-        rad;
-
-    function circle(x, y, r) {
-        context.moveTo(x + r, y);
-        context.arc(x, y, r, radians(0), radians(360));
-    }
-
-    // top left:
-    context.arc(
-        offset,
-        offset,
-        radius,
-        radians(-180),
-        radians(-90),
-        false
-    );
-    // top right:
-    context.arc(
-        w - offset,
-        offset,
-        radius,
-        radians(-90),
-        radians(-0),
-        false
-    );
-    // bottom right:
-    context.arc(
-        w - offset,
-        h - offset - radius,
-        radius,
-        radians(0),
-        radians(90),
-        false
-    );
-    if (!this.isThought) { // draw speech bubble hook
-        if (this.isPointingRight) {
-            context.lineTo(
-                offset + radius,
-                h - offset
-            );
-            context.lineTo(
-                radius / 2 + inset,
-                h - inset
-            );
-        } else { // pointing left
-            context.lineTo(
-                w - (radius / 2 + inset),
-                h - inset
-            );
-            context.lineTo(
-                w - (offset + radius),
-                h - offset
-            );
-        }
-    }
-    // bottom left:
-    context.arc(
-        offset,
-        h - offset - radius,
-        radius,
-        radians(90),
-        radians(180),
-        false
-    );
-    if (this.isThought) {
-        // close large bubble:
-        context.lineTo(
-            inset,
-            offset
-        );
-        // draw thought bubbles:
-        if (this.isPointingRight) {
-            // tip bubble:
-            rad = radius / 4;
-            circle(rad + inset, h - rad - inset, rad);
-            // middle bubble:
-            rad = radius / 3.2;
-            circle(rad * 2 + inset, h - rad - inset * 2, rad);
-            // top bubble:
-            rad = radius / 2.8;
-            circle(rad * 3 + inset * 2, h - rad - inset * 4, rad);
-        } else { // pointing left
-            // tip bubble:
-            rad = radius / 4;
-            circle(w - (rad + inset), h - rad - inset, rad);
-            // middle bubble:
-            rad = radius / 3.2;
-            circle(w - (rad * 2 + inset), h - rad - inset * 2, rad);
-            // top bubble:
-            rad = radius / 2.8;
-            circle(w - (rad * 3 + inset * 2), h - rad - inset * 4, rad);
-        }
-    }
-};
-
-// SpeechBubbleMorph shadow
-
-/*
-    only take the 'plain' image, so the box rounding and the
-    shadow doesn't become conflicted by embedded scrolling panes
-*/
-
-SpeechBubbleMorph.prototype.shadowImage = function (off, color) {
-    // fallback for Windows Chrome-Shadow bug
-    var fb, img, outline, sha, ctx,
-        offset = off || new Point(7, 7),
-        clr = color || new Color(0, 0, 0);
-    fb = this.extent();
-    img = this.image;
-    outline = newCanvas(fb);
-    ctx = outline.getContext('2d');
-    ctx.drawImage(img, 0, 0);
-    ctx.globalCompositeOperation = 'destination-out';
-    ctx.drawImage(
-        img,
-        -offset.x,
-        -offset.y
-    );
-    sha = newCanvas(fb);
-    ctx = sha.getContext('2d');
-    ctx.drawImage(outline, 0, 0);
-    ctx.globalCompositeOperation = 'source-atop';
-    ctx.fillStyle = clr.toString();
-    ctx.fillRect(0, 0, fb.x, fb.y);
-    return sha;
-};
-
-SpeechBubbleMorph.prototype.shadowImageBlurred = function (off, color) {
-    var fb, img, sha, ctx,
-        offset = off || new Point(7, 7),
-        blur = this.shadowBlur,
-        clr = color || new Color(0, 0, 0);
-    fb = this.extent().add(blur * 2);
-    img = this.image;
-    sha = newCanvas(fb);
-    ctx = sha.getContext('2d');
-    ctx.shadowOffsetX = offset.x;
-    ctx.shadowOffsetY = offset.y;
-    ctx.shadowBlur = blur;
-    ctx.shadowColor = clr.toString();
-    ctx.drawImage(
-        img,
-        blur - offset.x,
-        blur - offset.y
-    );
-    ctx.shadowOffsetX = 0;
-    ctx.shadowOffsetY = 0;
-    ctx.shadowBlur = 0;
-    ctx.globalCompositeOperation = 'destination-out';
-    ctx.drawImage(
-        img,
-        blur - offset.x,
-        blur - offset.y
-    );
-    return sha;
-};
-
-// SpeechBubbleMorph resizing
-
-SpeechBubbleMorph.prototype.fixLayout = function () {
-    this.removeShadow();
-    this.drawNew();
-    this.addShadow(new Point(2, 2), 80);
-};
-
-// CircleBoxMorph //////////////////////////////////////////////////////
-
-// I can be used for sliders
-
-var CircleBoxMorph;
-
-// CircleBoxMorph inherits from Morph:
-
-CircleBoxMorph.prototype = new Morph();
-CircleBoxMorph.prototype.constructor = CircleBoxMorph;
-CircleBoxMorph.uber = Morph.prototype;
-
-function CircleBoxMorph(orientation) {
-    this.init(orientation || 'vertical');
-}
-
-CircleBoxMorph.prototype.init = function (orientation) {
-    CircleBoxMorph.uber.init.call(this);
-    this.orientation = orientation;
-    this.autoOrient = true;
-    this.setExtent(new Point(20, 100));
-};
-
-CircleBoxMorph.prototype.autoOrientation = function () {
-    if (this.height() > this.width()) {
-        this.orientation = 'vertical';
-    } else {
-        this.orientation = 'horizontal';
-    }
-};
-
-CircleBoxMorph.prototype.drawNew = function () {
-    var radius, center1, center2, rect, points, x, y,
-        context, ext,
-        myself = this;
-
-    if (this.autoOrient) {
-        this.autoOrientation();
-    }
-    this.image = newCanvas(this.extent());
-    context = this.image.getContext('2d');
-
-    if (this.orientation === 'vertical') {
-        radius = this.width() / 2;
-        x = this.center().x;
-        center1 = new Point(x, this.top() + radius);
-        center2 = new Point(x, this.bottom() - radius);
-        rect = this.bounds.origin.add(new Point(0, radius)).corner(
-            this.bounds.corner.subtract(new Point(0, radius))
-        );
-    } else {
-        radius = this.height() / 2;
-        y = this.center().y;
-        center1 = new Point(this.left() + radius, y);
-        center2 = new Point(this.right() - radius, y);
-        rect = this.bounds.origin.add(new Point(radius, 0)).corner(
-            this.bounds.corner.subtract(new Point(radius, 0))
-        );
-    }
-    points = [ center1.subtract(this.bounds.origin),
-        center2.subtract(this.bounds.origin)];
-    points.forEach(function (center) {
-        context.fillStyle = myself.color.toString();
-        context.beginPath();
-        context.arc(
-            center.x,
-            center.y,
-            radius,
-            0,
-            2 * Math.PI,
-            false
-        );
-        context.closePath();
-        context.fill();
-    });
-    rect = rect.translateBy(this.bounds.origin.neg());
-    ext = rect.extent();
-    if (ext.x > 0 && ext.y > 0) {
-        context.fillRect(
-            rect.origin.x,
-            rect.origin.y,
-            rect.width(),
-            rect.height()
-        );
-    }
-};
-
-// CircleBoxMorph menu:
-
-CircleBoxMorph.prototype.developersMenu = function () {
-    var menu = CircleBoxMorph.uber.developersMenu.call(this);
-    menu.addLine();
-    if (this.orientation === 'vertical') {
-        menu.addItem(
-            "horizontal...",
-            'toggleOrientation',
-            'toggle the\norientation'
-        );
-    } else {
-        menu.addItem(
-            "vertical...",
-            'toggleOrientation',
-            'toggle the\norientation'
-        );
-    }
-    return menu;
-};
-
-CircleBoxMorph.prototype.toggleOrientation = function () {
-    var center = this.center();
-    this.changed();
-    if (this.orientation === 'vertical') {
-        this.orientation = 'horizontal';
-    } else {
-        this.orientation = 'vertical';
-    }
-    this.silentSetExtent(new Point(this.height(), this.width()));
-    this.setCenter(center);
-    this.drawNew();
-    this.changed();
-};
-
 // SliderButtonMorph ///////////////////////////////////////////////////
 
 var SliderButtonMorph;
