+++ src/utils.js
@@ -1,0 +1,190 @@
+// Global Functions ////////////////////////////////////////////////////
+
+function nop() {
+    // do explicitly nothing
+    return null;
+}
+
+function localize(string) {
+    // override this function with custom localizations
+    return string;
+}
+
+function isNil(thing) {
+    return thing === undefined || thing === null;
+}
+
+function contains(list, element) {
+    // answer true if element is a member of list
+    return list.some(function (any) {
+        return any === element;
+    });
+}
+
+function detect(list, predicate) {
+    // answer the first element of list for which predicate evaluates
+    // true, otherwise answer null
+    var i, size = list.length;
+    for (i = 0; i < size; i += 1) {
+        if (predicate.call(null, list[i])) {
+            return list[i];
+        }
+    }
+    return null;
+}
+
+function sizeOf(object) {
+    // answer the number of own properties
+    var size = 0, key;
+    for (key in object) {
+        if (Object.prototype.hasOwnProperty.call(object, key)) {
+            size += 1;
+        }
+    }
+    return size;
+}
+
+function isString(target) {
+    return typeof target === 'string' || target instanceof String;
+}
+
+function isObject(target) {
+    return target !== null &&
+        (typeof target === 'object' || target instanceof Object);
+}
+
+function radians(degrees) {
+    return degrees * Math.PI / 180;
+}
+
+function degrees(radians) {
+    return radians * 180 / Math.PI;
+}
+
+function fontHeight(height) {
+    var minHeight = Math.max(height, MorphicPreferences.minimumFontHeight);
+    return minHeight * 1.2; // assuming 1/5 font size for ascenders
+}
+
+function newCanvas(extentPoint) {
+    // answer a new empty instance of Canvas, don't display anywhere
+    var canvas, ext;
+    ext = extentPoint || {x: 0, y: 0};
+    canvas = document.createElement('canvas');
+    canvas.width = ext.x;
+    canvas.height = ext.y;
+    return canvas;
+}
+
+function getMinimumFontHeight() {
+    // answer the height of the smallest font renderable in pixels
+    var str = 'I',
+        size = 50,
+        canvas = document.createElement('canvas'),
+        ctx,
+        maxX,
+        data,
+        x,
+        y;
+    canvas.width = size;
+    canvas.height = size;
+    ctx = canvas.getContext('2d');
+    ctx.font = '1px serif';
+    maxX = ctx.measureText(str).width;
+    ctx.fillStyle = 'black';
+    ctx.textBaseline = 'bottom';
+    ctx.fillText(str, 0, size);
+    for (y = 0; y < size; y += 1) {
+        for (x = 0; x < maxX; x += 1) {
+            data = ctx.getImageData(x, y, 1, 1);
+            if (data.data[3] !== 0) {
+                return size - y + 1;
+            }
+        }
+    }
+    return 0;
+}
+
+function getBlurredShadowSupport() {
+    // check for Chrome issue 90001
+    // http://code.google.com/p/chromium/issues/detail?id=90001
+    var source, target, ctx;
+    source = document.createElement('canvas');
+    source.width = 10;
+    source.height = 10;
+    ctx = source.getContext('2d');
+    ctx.fillStyle = 'rgb(255, 0, 0)';
+    ctx.beginPath();
+    ctx.arc(5, 5, 5, 0, Math.PI * 2, true);
+    ctx.closePath();
+    ctx.fill();
+    target = document.createElement('canvas');
+    target.width = 10;
+    target.height = 10;
+    ctx = target.getContext('2d');
+    ctx.shadowBlur = 10;
+    ctx.shadowColor = 'rgba(0, 0, 255, 1)';
+    ctx.drawImage(source, 0, 0);
+    return ctx.getImageData(0, 0, 1, 1).data[3] ? true : false;
+}
+
+function getDocumentPositionOf(aDOMelement) {
+    // answer the absolute coordinates of a DOM element in the document
+    var pos, offsetParent;
+    if (aDOMelement === null) {
+        return {x: 0, y: 0};
+    }
+    pos = {x: aDOMelement.offsetLeft, y: aDOMelement.offsetTop};
+    offsetParent = aDOMelement.offsetParent;
+    while (offsetParent !== null) {
+        pos.x += offsetParent.offsetLeft;
+        pos.y += offsetParent.offsetTop;
+        if (offsetParent !== document.body &&
+                offsetParent !== document.documentElement) {
+            pos.x -= offsetParent.scrollLeft;
+            pos.y -= offsetParent.scrollTop;
+        }
+        offsetParent = offsetParent.offsetParent;
+    }
+    return pos;
+}
+
+function clone(target) {
+    // answer a new instance of target's type
+    if (typeof target === 'object') {
+        var Clone = function () {nop(); };
+        Clone.prototype = target;
+        return new Clone();
+    }
+    return target;
+}
+
+function copy(target) {
+    // answer a shallow copy of target
+    var value, c, property;
+
+    if (typeof target !== 'object') {
+        return target;
+    }
+    value = target.valueOf();
+    if (target !== value) {
+        return new target.constructor(value);
+    }
+    if (target instanceof target.constructor &&
+            target.constructor !== Object) {
+        c = clone(target.constructor.prototype);
+        for (property in target) {
+            if (Object.prototype.hasOwnProperty.call(target, property)) {
+                c[property] = target[property];
+            }
+        }
+    } else {
+        c = {};
+        for (property in target) {
+            if (!c[property]) {
+                c[property] = target[property];
+            }
+        }
+    }
+    return c;
+}
